FROMLIST: arm64: Switch to adrp for loading the stub vectors
The hyp stub vectors are currently loaded using adr. This
instruction has a +/- 1MB range for the loading address. If
the alignment for sections is changed the address may be more
than 1MB away, resulting in reclocation errors. Switch to using
adrp for getting the address to ensure we aren't affected by the
location of the __hyp_stub_vectors.

Acked-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Acked-by: Marc Zyngier <marc.zyngier@arm.com>
Tested-by: Mark Rutland <mark.rutland@arm.com>
Tested-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>

BUG=chromium:436130
TEST=ryu boot, lkdtm WRITE_RO and WRITE_KERN trigger Oopses

(v6: https://patchwork.kernel.org/patch/5358161/)
Signed-off-by: Kees Cook <keescook@chromium.org>
Change-Id: Ieb671f46f0a9c1c12c2f430b19c8d807bf19cb16
Reviewed-on: https://chromium-review.googlesource.com/231765
Reviewed-by: Benson Leung <bleung@chromium.org>