CHROMIUM: drm/tegra: Add falcon helper library
Add a set of falcon helper routines for use by the tegradrm client drivers
of the various falcon-based engines.

The falcon is a microcontroller that acts as a frontend for the rest of a
particular Tegra engine.  In order to properly utilize these engines, the
frontend must be booted before pushing any commands.

Based on work by Arto Merilainen <amerilainen@nvidia.com>.

BUG=chrome-os-partner:39847
TEST=Built and booted on smaug

Change-Id: I7d46214b4cdfee5e01d200ba80c692123013eec0
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/269428
Reviewed-by: Sean Paul <seanpaul@google.com>