UPSTREAM: ath10k: add register access debugfs interface
Debugfs files reg_addr and reg_val are used for reading and writing to the
firmware (target) registers. reg_addr contains the address to be accessed,
which also needs to be set first, and reg_value is when used for reading and
writing the actual value in ASCII.

To read a value from the firmware register 0x100000:

# echo 0x100000 > reg_addr
# cat reg_value
0x00100000:0x000002d3

To write value 0x2400 to address 0x100000:

# echo 0x100000 > reg_addr
# echo  0x2400 > reg_value
#

Signed-off-by: Yanbo Li <yanbol@qti.qualcomm.com>
Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>
(cherry picked from commit 077a380447160d94031c51e863ca4c53d2c74730)
(source: https://github.com/kvalo/ath/tree/ath-qca)

BUG=chrome-os-partner:35262
TEST=build/boot on storm suceeded. Setup Storm board as AP using hostapd and
connected to the board using another device. Device was able to connect to
the internet and load multiple websites.

Change-Id: I5a1e4f4395b27cbd0127fa901a45e0fb0e54f268
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/241631
Reviewed-by: Toshi Kikuchi <toshik@chromium.org>