FROMLIST: clk: qcom: Add simple regmap based muxes
Add support for muxes that use regmap instead of readl/writel
directly. We don't support as many features as clk-mux.c, but
this is good enough to support getting and setting parents.
Adding a table based lookup can be added in the future if needed.

BUG=chrome-os-partner:27830
TEST=Ran aplay with 48KHz file
     % aplay -D 'hw:0,0' -v --dump-hw-params /usr/share/kcw/test.wav

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
(am from https://patchwork.kernel.org/patch/5383441/)
Signed-off-by: Kenneth Westfield <kwestfie@codeaurora.org>
Change-Id: I6ed5b1ce0c9f7b221c508c540c7ebcac9346d08b
Reviewed-on: https://chromium-review.googlesource.com/233940
Reviewed-by: Dylan Reid <dgreid@chromium.org>