mm: fix anon_vma_clone() error treatment
Andrew Morton noticed that the error return from anon_vma_clone() was
being dropped and replaced with -ENOMEM (which is not itself a bug because
the only error return value from anon_vma_clone() is -ENOMEM).

I did an audit of callers of anon_vma_clone() and discovered an actual bug
where the error return was being lost.  In __split_vma(), between Linux
3.11 and 3.12 the code was changed so the err variable is used before the
call to anon_vma_clone() and the default initial value of -ENOMEM is
overwritten.  So a failure of anon_vma_clone() will return success since
err at this point is now zero.

Below is a patch which fixes this bug and also propagates the error return
value from anon_vma_clone() in all cases.

Signed-off-by: Daniel Forrest <dan.forrest@ssec.wisc.edu>
Cc: Konstantin Khlebnikov <koct9i@gmail.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Rik van Riel <riel@redhat.com>
Cc: Tim Hartrick <tim@edgecast.com>
Cc: Hugh Dickins <hughd@google.com>
Cc: Michel Lespinasse <walken@google.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>