BACKPORT: PM / OPP Introduce APIs to remove OPPs
OPPs are created statically (from DT) or dynamically. Currently we don't free
OPPs that are created statically, when the module unloads. And so if the module
is inserted back again, we get warning for duplicate OPPs as the same were
already present.

Also, there might be a need to remove dynamic OPPs in future and so API for that
is also added.

This patch adds helper APIs to remove/free existing static and dynamic OPPs.

Because the OPPs are used both under RCU and SRCU, we have to wait for grace
period of both. And so are using kfree_rcu() from within call_srcu().

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 129eec55df6ab1b5ecdd89fd7db7a2cd103200b5)
Signed-off-by: Pi-Cheng Chen <pi-cheng.chen@mediatek.com>

[pi-cheng.chen: this patch has a trivial conflict with
 chromeos-3.18 commit [0] which added of_init_opp_table_named()
 and converted of_init_opp_table() to call it.
 [0] 6aff17c221456a710682af44a42147a8e60a96a9
 CHROMIUM: PM / OPP: introduce of_init_opp_table_named
]

BUG=chrome-os-partner:39501
TEST=build and boot oak-rev2 and test each OPP of two clusters

Change-Id: Ib3730bf6c90fbc9faee219f9799ab0a2f1480c62
Reviewed-on: https://chromium-review.googlesource.com/278130
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Trybot-Ready: Daniel Kurtz <djkurtz@chromium.org>