UPSTREAM: xhci: clear extra bits from slot context when setting max exit latency
If we need to change the max exit latency with a Evaluate Context
command, we copy the old output slot context and use it as input
context for the command. This also copies the dev_state bits which
are supposed to be zero in the input slot context.

Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Link: git://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git
br: usb-next
(cherry picked from commit 4801d4ea559440b7d63babdac1b1b43a13ac7562)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=None
TEST=with the whole series applied, the USB subsystem is more stable.
TEST=there are no regressions or file system related errors observed
TEST=on SKL based systems.

Change-Id: I9517e2b05276d6f9d8ec718361c2f580620ba6a7
Reviewed-on: https://chromium-review.googlesource.com/272756
Tested-by: Philip Hanson <philip.hanson@intel.com>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Vedang Patel <vedang.patel@intel.com>