CHROMIUM: host1x: Implement engine reset
Currently host1x cleans up the cdma and makes the additional
increments. However, the engine is not reset and as a result the
engine is likely unusable on the subsequent jobs.

This patch implements a callback to reset module after timeout. The
actual reset is then done in the driver code.

BUG=chrome-os-partner:39847
TEST=Built and booted on smaug

Change-Id: I53de49f0dd055af254e22c9893b83121b3fcf10c
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/269422
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>