iser-target: Move protection information handling to a routine
Centralize the PI context handling code. This is a pre-step
for dynamic allocation of PI context (When we put the allocation
in this routine it won't be weird).

Also, keep ib_sge array in isert_rdma_wr to avoid abusing the
stack too much by passing those as function arguments.

Cc: <stable@vger.kernel.org> # 3.15+
Signed-off-by: Sagi Grimberg <sagig@mellanox.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>