UPSTREAM: tpm/tpm_i2c_stm_st33: Fix potential bug in tpm_stm_i2c_send
When sending data in tpm_stm_i2c_send, each loop iteration send buf.
Send buf + i instead as the goal of this for loop is to send a number
of byte from buf that fit in burstcnt. Once those byte are sent, we are
supposed to send the next ones.

The driver was working because the burstcount value returns always the maximum size for a TPM
command or response. (0x800 for a command and 0x400 for a response).

Cc: stable@vger.kernel.org

Reviewed-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>
Signed-off-by: Christophe Ricard <christophe-h.ricard@st.com>
Signed-off-by: Peter Huewe <peterhuewe@gmx.de>
(cherry picked from commit 1ba3b0b6f218072afe8372d12f1b6bf26a26008e)

BUG=chrome-os-partner:47234
TEST=verify no regression in TPM 1.2 systems

Change-Id: Ia8df6a6780b4166db87aa7b6389d4cf255e3685e
Signed-off-by: Bryan Freed <bfreed@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/316540
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>