FROMLIST: ARM: rockchip: Convert resume code to C
This CL introduces the concept of "embedded blobs" for Rockchip, which
allows you to compile some C code into a binary blob that is linked
into the kernel.  This binary blob is self contained and is intended
to run in cases where SDRAM (and thus the rest of Linux) isn't
available.  Resume is a prime candiate for this as is the planned
DDRFreq code.

We convert the existing assembly resume code into C as proof that this
works and to prepare for linking in SDRAM reinit code.

At the moment suspend/resume is only implemented for rk3288 and I'm
told that suspend/resume on older Rockchip variants is very limited
(can't shut off the ARM?), so I haven't tried to make this generic.
When we have more than one Rockchip SoC doing S2R we can adjust.

Conflicts:
    arch/arm/mach-rockchip/Makefile
...due to local domain code

BUG=chrome-os-partner:32377
TEST=Suspend/resume still works

Change-Id: Icdf88ffbb4890d7fd722ba3a5cfe41799978136e
Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/5414941/)
Reviewed-on: https://chromium-review.googlesource.com/225492
Reviewed-by: Julius Werner <jwerner@chromium.org>
Tested-by: Chris Zhong <zyw@rock-chips.com>