CHROMIUM: ARM: dts: Add rise and fall times for veyron
These are based on measurements on pinky rev2 with a bit of margin
added.  I didn't put in real values for i2c3 since it was measured
without a touchscreen in place.  I assumed that the fall time will
still be fast there, though.  I also put in max values (including max
fall time) for i2c5 since that goes to HDMI and you could imagine a
user attaching all sorts of things.

BUG=chrome-os-partner:34207
TEST=Measure with scope

Change-Id: Ic565f8b2996b2e8d786d2f1fdfa68a15a1ed28ba
Signed-off-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: cyw <cyw@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/232775
Reviewed-by: Alexandru Stan <amstan@chromium.org>