CHROMIUM: FROMLIST: ASoC: rockchip: i2s: add support for grabbing output clock to codec
We need to claim the clock which is driving the codec so that when we enable
clock gating, we continue to clock the codec when needed.  I make this an
optional clock since there might be some applications where we don't need it
but can still use the I2S block.

BUG=chrome-os-partner:33886
TEST=audio works on pinky rev2

Change-Id: I5a951e289a6f2cc321221c27885a7b262ae3b908
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Signed-off-by: Jianqun Xu <jay.xu@rock-chips.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/5427131/)
Reviewed-on: https://chromium-review.googlesource.com/230347