CHROMIUM: rockchip: set the gpu/core power domain power delay time
When core or GPU domain powers up or down the hardware will not signal
completion until the delay specified in PMU_(CORE|GPU)_PWRDWN_CNT and
PMU_(CORE|GPU)_PWRUP_CNT registers passes. The default values in these
registers are equivalent to 1 msec, which is too long. According to the
TRM the maximum transition time for all domains is less than 900 nsecs
so we can safely program the both counters to 1 usec.

BUG=chrome-os-partner:33950
TEST=Boot to UI;
We can get the registers value.
$io -r -4 0xff730034 ----> 0x18
$io -r -4 0xff730038 ----> 0x18
$io -r -4 oxff73003c ----> 0x18
$io -r -4 oxff730040 ----> 0x18

Change-Id: I11293f04a19f57a2cc0414f0b04ce4840f6bea30
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/231629
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>