CHROMIUM: pinctrl: tegra-xusb: Add suspend/resume support
The XUSB pad-controller loses pad mux state and must be re-enabled
after resume.  The suspend/resume operations force the pad controller
into and out of the sleep state (supplied by the device-tree) across
suspend/resume.  It's expected that this sleep state will be the
same as the runtime state on Tegra124/Tegra132.

APIs are also added for use by the xHCI driver during the host
powergate/unpowergate procedure.

BUG=chrome-os-partner:33677
TEST=Suspend/resume on Ryu; USB still works.

Change-Id: I4f6e51c7d0f53bfbfc9eef691e220752fbb6e514
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/233293
Reviewed-by: Benson Leung <bleung@chromium.org>