[temporary] multi_v7_defconfig: turn on list debugging
I'm trying to get to the bottom of the regression that caused
the boot regression on Capri as in
http://arm-soc.lixom.net/bootlogs/arm-soc/v3.18-rc7-1139-ge2e56da/capri-arm-multi_v7_defconfig.html

I have figured out that the klist_devices list in bdi_class
is corrupted at the point when the loopback driver gets loaded.

I suspected MTD as the root cause, so my first attempt was to
disable MTD and enable list debugging, which made the problem
disappear. I then enabled MTD again and added more debugging
printks, but that also managed to hide the bug, so now I'm
trying with just list debugging.

Unfortunately, I can't test this locally, only by uploading
to the arm-soc tree and watching the boot bot output.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>