UPSTREAM: mmc: sdhci: Disable re-tuning for HS400
Re-tuning for HS400 mode must be done in HS200
mode. Currently there is no support for that.
That needs to be reflected in the code.
Specifically, if tuning is executed in HS400 mode
then return an error, and do not start the
tuning timer if HS200 tuning is being done prior
to switching to HS400.

Note that periodic re-tuning is not expected
to be needed for HS400 but re-tuning is still
needed after the host controller has lost power.
In the case of suspend/resume that is not necessary
because the card is fully re-initialised. That
just leaves runtime suspend/resume with no support
for HS400 re-tuning.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
(cherry picked from commit b5540ce1512eede3bed68ab1e9949df9ad556091)
Signed-off-by: Marc Herbert <marc.herbert@intel.com>

BUG=chrome-os-partner:40410
TEST=With this whole HS400 series applied:
TEST=on a system with HS400 controller(s), it/they initialize with the
  following log message:
  INFO kernel: [  103.601747] mmc0: new HS400 MMC card at address 0001
TEST=run SD/eMMC test plan: higher eMMC throughput after change, and no
  other change.
TEST=same on another system with no HS400: no observable change at all.

Change-Id: I0a26f0aedc04d6e4f1a516e8a7ea9769dcf4ff9d
Reviewed-on: https://chromium-review.googlesource.com/276393
Commit-Ready: Marc Herbert <marc.herbert@intel.com>
Tested-by: Philip Hanson <philip.hanson@intel.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>