mmc: dw_mmc: adjust scope of DW_MCI_QUIRK_BROKEN_DTO to fix emmc hangs
Seems like the BROKEN_DTO also affects EVENT_DATA_COMPLETE, sometimes we don't
see that (but we do see EVENT_XFER_COMPLETE), leave the state machine and never
come back to it to timeout. I have this happening on emmc specifically, haven't
seen it on sdmmc yet.

The fix is to also start the timer when we don't see a EVENT_DATA_COMPLETE.

More details for how I debugged this can be found at:
https://chromium-review.googlesource.com/#/c/233691/

BUG=chrome-os-partner:32740
TEST=emmc will not freeze anymore at hs200 (NOTE: freezes were rare when it did)

Change-Id: I83d347004fc54a577f669826003ad287fa7dfb59
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/233694
Reviewed-by: Doug Anderson <dianders@chromium.org>