CHROMIUM: soc: tegra132: fix build error with PM_SLEEP
drivers/built-in.o: In function `start_ns':
	kernel/v3.14/drivers/soc/tegra/sleep-tegra132.S:213:(.head.text+0x74):
	relocation truncated to fit: R_AARCH64_CONDBR19 against symbol
	`secondary_holding_pen' defined in .text
	section in arch/arm64/kernel/head.o

This patch fixes the link time error when tegra_resume function used the
branch instruction to reach the secondary_holding_pen in the kernel.
Replace that by using the PC-relative address generation instruction.
And we don't need to insert tegra_resume into __HEAD section, just leave
it in the normal text section.

BUG=chrome-os-partner:31211
TEST=build and boot ok with PM_SLEEP on Ryu,
     LP0 is OK as well.

Change-Id: Iaaab1056ee1e38d785965f8e936b9a874100ad04
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Signed-off-by: Dylan Reid <dgreid@chromium.org>