drm/tegra: sor: Implement ->atomic_check()
The implementation of the ->atomic_check() callback precomputes all
parameters to check if the given configuration can be applied. If so the
precomputed values are stored in the atomic state object for the encoder
and applied during modeset. In that way the modeset no longer needs to
perform any checking but simply program values into registers.

Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit 82f1511c35643575e6f28013a3bb348e35676491)
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>