CHROMIUM: serial: tegra: handle race condition on uart rx side
Skip the uart rx dma callback handling if dma status is in progress to avoid
race condition.  This prevents an extra/erroneous copy of data.

The race condition happens when the UART RX DMA completion interrupt occurs and
the apbdma driver is executing tasklet() in which serial tegra completion
callback is not called yet and meanwhile UART RX interrupt occurs
simultaneously.

BUG=chrome-os-partner:39890
TEST=/dev/ttyHS2 and /dev/ttyHS3 appear

Change-Id: I32d10c1719853b4e85d9aa30eb423d55f03ebd9a
Signed-off-by: Shardar Shariff Md <smohammed@nvidia.com>
Signed-off-by: Christopher Freeman <cfreeman@nvidia.com>
Signed-off-by: David Riley <davidriley@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/269781
Reviewed-by: Dmitry Shmidt <dimitrysh@google.com>