FIXUP: FROMLIST: i2c: rk3x: fix bug that cause measured high_ns doesn't meet I2C specification
This is a squash of a revert of:
  d41d56e FROMLIST: i2c: rk3x: fix bug that cause measured high_ns doesn't meet I2C spec
...and a new pick from the list.

---

The number of clock cycles to be written into the CLKDIV register
that determines the I2C clk high phase includes the rise time.
So to meet the timing requirements defined in the I2C specification
which defines the minimal time SCL has to be high, the rise time
has to taken into account. The same applies to the low phase with
falling time.

In my test on RK3288-Pink2 board, which is not an upstream board yet,
if external pull-up resistor is 4.7K, rise_ns is about 700ns.
So the measured high_ns is about 3900ns, which is less than 4000ns
(the minimum high_ns in I2C specification for Standard-mode).

To fix this bug min_low_ns should include fall time and min_high_ns
should include rise time.

This patch merged the patch from chromium project which can get the
rise and fall times for signals from the device tree. This allows us
to more accurately calculate timings. see:
https://chromium-review.googlesource.com/#/c/232774/

BUG=chrome-os-partner:34207
TEST=i2c still works

Change-Id: I36743ee44b22aeced1801c913a6b69457c619b38
Signed-off-by: Addy Ke <addy.ke@rock-chips.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/5475331/)
Reviewed-on: https://chromium-review.googlesource.com/234637
Reviewed-by: Alexandru Stan <amstan@chromium.org>