CHROMIUM: drivers: ar8216: read vlan init from DT
Each platform has a different way of initializing the ethernet switch,
which mainly depends on its ethernet layout. So we'll put this
information in device-tree. It will be applied both at init, and when
calling "swconfig dev switch0 set reset 1".

Port based VLANs can now be configured using the following format in the
device tree, in the switch device:
	qca,ar8327-vlans = <
		VLANID1 PORTMASK1
		VLANID2 PORTMASK2
		...
	>;

Where VLANID is the ID used in the switch device to identify it, and
PORTMASK is the bit mask of all the ports we want in this particular
VLAN.

Typically, if I want to create a VLAN1 with ports 0 and 1, I'd have a
line:	0x1 0x3

BUG=chrome-os-partner:33096
TEST=none

Change-Id: If500f70f0df80d1aaa7b9545acabeecb664227da
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/235501
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Toshi Kikuchi <toshik@chromium.org>
Commit-Queue: Toshi Kikuchi <toshik@chromium.org>
Tested-by: Toshi Kikuchi <toshik@chromium.org>