UPSTREAM: ASoC: dapm: Don't mark MICBIAS widgets as auto non-connected
The connected flag of a widget only affects widgets that are either a source
or a sink. The MICBIAS widget is a simple pass-through widget though and
hence its behavior is the same regardless of whether the connected flag is
set or not.

Hence there is not much point in trying to automatically mark MICBIAS
widgets as non-connected, so just remove it.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 768c0564439bce179e3f0f2ffafe8dcab6a08f40)
BUG=chrome-os-partner:40036
TEST=Audio playback works on test device

Change-Id: Iadff2af42b9ffe81de26c1687704db3645f691c6
Reviewed-on: https://chromium-review.googlesource.com/270937
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>