FROMLIST: pinctrl: rockchip: Avoid losing interrupts when supporting both edges
I was seeing cases where I was losing interrupts when inserting and
removing SD cards.  Sometimes the card would get "stuck" in the
inserted state.

I believe that the problem was related to the code to handle the case
where we needed both rising and falling edges.  This code would
disable the interrupt as the polarity was switched.  If an interrupt
came at the wrong time it could be lost.

We'll match what the gpio-dwapb.c driver does upstream and change the
interrupt polarity without disabling things.

BUG=chrome-os-partner:34752
TEST=Eject SD card a lot; it's always detected

Change-Id: I56e4d0ee84d7700cef7944ceb899f5ca55cefe38
Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/5528731/)
Reviewed-on: https://chromium-review.googlesource.com/236701
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>