Revert "Revert "CHROMIUM: arm: dts: whirlwind: enable the third pcie slot""
This reverts commit 56f53b08151fd737a5be8768d353623f07779d61.

We had disabled the third pcie slot with the previous commit because
a kernel bug in lib/genalloc.c caused ath10k driver to crash if three
ath10k driver instances are enabled at the same time. We have fixed the
kernel bug so we can re-enable the third pcie slot.

BUG=chrome-os-partner:33870
TEST=boot whirlwind image and ath10k driver won't panic

Change-Id: I913541b025a7d793354484d82087b93c870aab53
Signed-off-by: Toshi Kikuchi <toshik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/237304
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>