CHROMIUM: rockchip: pll: Actually use pll clk flags
The pll struct has a member that stores the clk flags for that pll, but it is
ignored when the pll is actually registered. This fixes that and removes the
CLK_GET_RATE_NOCACHE flag from being added by the pll macro. It also adds it
back in for the dpll since it actually needs that flag for ddr freq to work.

BUG=chrome-os-partner:31440
TEST=boot jerry

Change-Id: I9afb808158556f964cf4089bc4184259689e09a6
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/239310
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>