drm/nouveau/nvif: namespace of nvkm accessors (no binary change)
NVKM is having it's namespace switched to nvkm_, which will conflict
with these functions (which are workarounds for the fact that as of
yet, we still aren't able to split DRM and NVKM completely).

A comparison of objdump disassemblies proves no code changes.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>
(cherry picked from commit 989aa5b76ad2af7653353cf01bdebec2ba9436aa)
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>