FROMLIST: mmc: dw_mmc: rockchip: Add DW_MCI_QUIRK_RETRY_DELAY
We've seen at least one card that can get confused during all the
errors generated during tuning on rk3288.  These errors seem to go
away with DW_MCI_QUIRK_RETRY_DELAY.  That quirk is documented to be
for 2.11a and on rk3288 I believe we have 2.70a, but it seems to be
the same type of problem.

Specifically it looks like in certain error conditions (I saw this
with Response CRC errors) that data keeps showing up in the FIFO even
after the error is reported and the CD (command done) bit is set.  If
we don't wait for this data to finish transferring then it confuses
the next transaction.  In the specific failure case I ran into I found
that I could monitor the data_state_mc_busy bit and wait for it to
clear, but in other failure cases this bit was stuck at busy when we
saw an error.  Hence a generic big delay seems like the only option.

I have seen this on rk3288, but adding it as a quirk for all Rockchip
IP for now since most previous quirks have shown up on older IP as
well.

BUG=chrome-os-partner:32740
TEST=Reboot tests with problematic card now pass

Change-Id: I591926cf498914818b7bdfc4a2eb114d5f5da5b9
Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/5623071/)
Reviewed-on: https://chromium-review.googlesource.com/240205
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>