CHROMIUM: Change default value of the PSR source parameter 'idleframes' to 3
We are hearing reports of screen flickering from left to right. We are
currently conjecturing that this is PSR related and increasing idleframes
to 3 may help. Unfortunately we do not have a good repro for the issue
and therefore it will be hard to prove or disprove this.

BUG=chrome-os-partner:34320
TEST=on Samus run " cat /sys/kernel/debug/dri/0/i915_edp_psr_status"
and verify that the Performance_Counter field increments when the screen
is idle.

Change-Id: I1769096f8473978f27d54dba5a33351879bed9d7
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/241171
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>