FROMLIST: PM / QoS: Make it possible to expose device latency tolerance to userspace
Typically when a device is created the bus core it belongs to (for example
PCI) does not know if the device supports things like latency tolerance.
This is left to the driver that binds to the device in question. However,
at that time the device has already been created and there is no way to set
its dev->power.set_latency_tolerance anymore.

So follow what has been done for other PM QoS attributes as well and allow
drivers to expose and hide latency tolerance from userspace, if the device
supports it.

Acked-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Link: https://patchwork.kernel.org/patch/6607491/
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

Conflicts:
	(Adjusting to contextual changes)
	drivers/base/power/power.h
	drivers/base/power/qos.c
	include/linux/pm_qos.h

BUG=chrome-os-partner:40857
TEST=In the output of lspci -v, intel-lpss name gets attached to most
   "Signal processing controller" sections.
TEST=Two new lines "Synopsys Designware I2C adapter" appear in:
 - cat /sys/bus/i2c/devices/*/name
 - and in the output "i2cdetect -l" (i2cdetect on "test" image at least)
TEST=Two new "i2c_designware" lines appear in /proc/interrupts

Change-Id: Ib20abd8a8d48b0ef16b97bb5d54b1dca678078a4
Reviewed-on: https://chromium-review.googlesource.com/282871
Reviewed-by: Joe Konno <joe.konno@linux.intel.com>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Tested-by: Philip Hanson <philip.hanson@intel.com>
Commit-Queue: Vedang Patel <vedang.patel@intel.com>