CHROMIUM: usb: Retry port status check on resume to work around RH bugs
The EHCI controller on the RK3288 SoC is violating basic parts of the
USB spec and thereby unable to properly resume a suspended port. It does
not start SOF generation within 3ms of finishing resume signaling, so
the attached device will drop of the bus again. This is a particular
problem with runtime PM, where accessing the device will trigger a
resume that immediately makes it unavailabe (and reenumerate with a new
handle).

Thankfully, the persist feature is generally able to work around stuff
like that. Unfortunately, it doesn't quite work in this particular case
because the controller will turn off the CurrentConnectStatus bit for an
instant while the device is reconnecting, which causes the kernel to
conclude that it permanently disappeared. This patch adds a tiny retry
mechanism to the core port resume code which will catch this case and
shouldn't have any notable impact on other controllers.

BUG=chrome-os-partner:34101
TEST=Boot Jerry, run 'echo 1 > /sys/bus/usb/devices/3-1/power/persist',
open Camera app and note how it doesn't display an error message.

Change-Id: I4b63059debc4278609c63df4e9f3823c86ad6cd1
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/234326
Reviewed-by: Douglas Anderson <dianders@chromium.org>

Conflicts:
	drivers/usb/core/hub.c
[filbranden: 3.18 rebase. Resolved complex conflicts.]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>