input: pmic8xxx-pwrkey: send correct key status in selective suspend state
When the device is in selective suspend state and PWR key is pressed or
released, the ISR gets called but before the key status is communicated to
the userspace,the device goes back into suspend state because the USB is
in suspend as well. In cases where if we want keypad release key event
trigger shutdown to happen, the shutdown is delayed until some other
interrupt is received or selective suspend is removed. This solution uses
a wakelock for over 1 sec which allows the user space to read the key status
correctly and hence the shutdown won't be delayed.

Change-Id: Ib06343b986c040b6bceb83dfe1d7f907149735df
Signed-off-by: Karthik Jadala <karthikjk@codeaurora.org>