CHROMIUM: ARM: dts: Hack the BRCM "bt_dev_wake" to be low in S3, high otherwise
On newer versions of veyron boards we'll hook GPIO4_D2 (AKA GPIO4_26)
up to bt_dev_wake.  We don't have an LPM driver yet so we can't
achieve power savings when we're in S3, but for now we'll use a
pinctrl hog to get the pin low in S3 and high in S0.

BUG=chrome-os-partner:33382
TEST=Look at power at suspend time

Change-Id: If77749c09278ebff5154fe0358198916c7574708
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/242574
Reviewed-by: Olof Johansson <olofj@chromium.org>