CHROMIUM: ARM: dts: Hack the BRCM "bt_dev_wake" to be low in S3, high otherwise
NOTE:
  This is take 2 of the CL: the first was reverted because it depended
  on another CL that was reverted.  Going back to the old version of
  CL:242574 version 2 which doesn't depend on the CL that was
  reverted.

On newer versions of veyron boards we'll hook GPIO4_D2 (AKA GPIO4_26)
up to bt_dev_wake.  We don't have an LPM driver yet so we can't
achieve power savings when we're in S3, but for now we'll use a
pinctrl hog to get the pin low in S3 and high in S0.

BUG=chrome-os-partner:33382, chrome-os-partner:37022
TEST=Look at power at suspend time

Change-Id: I4c508850df521f6d5bc5f04382382c64ca4d7ac7
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/252841
Reviewed-by: Alexandru Stan <amstan@chromium.org>