CHROMIUM: cros-ec: Fix usage of cros_ec_cmd_xfer
cros_ec_cmd_xfer returns success status if the command transport
completes successfully, but the execution result is incorrectly ignored.
In many cases, the execution result is assumed to be successful, leading
to ignored errors and operating on uninitialized data.

Add a new function cros_ec_cmd_xfer_status that checks the execution result
and returns error status if it's non-success. Convert most uses in the
codebase (those which don't already decode the result) to use this new
function.

BUG=chromium:451105
TEST=Manual on Samus. Verify PD AU and PD charger drivers still function
correctly.

Change-Id: Ic35d0cd6358435f027d21a562d60165a3b90d6c7
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/242722
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>

Conflicts:
	drivers/i2c/busses/i2c-cros-ec-tunnel.c
	drivers/iio/accel/cros_ec_accel.c
	drivers/input/keyboard/cros_ec_keyb.c
	drivers/mfd/cros_ec.c
	drivers/mfd/cros_ec_pd_update.c
	drivers/power/cros_ec-charger.c
	drivers/power/cros_usbpd-charger.c
	drivers/regulator/cros_ec-tps65090.c
	include/linux/mfd/cros_ec.h