CHROMIUM: i2c: img-scb: Fix I2C data reads
Currently, single-byte I2C reads always time out in Pistachio.
Snooping the I2C bus shows the transaction is correct, though.

Tracing the code showed that the FIFO-filling interrupt is never
triggered, which is required for the automatic-mode read to succeed.

In order to fix this, unmask the stop-bit detect interrupt in automatic
mode and use it to drain the FIFO and signal transaction completion.

On TZ1090 SoCs, the FIFO filling interrupt would be received first and,
after reading all the data, the mode would be switched to MODE_WAITSTOP
which uses a differente interrupt mask. In other words, TZ1090 SoCs
should be unaffected by this change.

BUG=chrome-os-partner:35348
TEST=Tested on Pistachio BuB

Change-Id: Idd62464a5fa5b7cefa054b27271d1ede595c2621
Signed-off-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/243917
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>