CHROMIUM: iwl7000: pcie: disable the SCD_BASE_ADDR when we resume from WoWLAN
The base address of the scheduler in the device's memory
(SRAM) comes from two different sources. The periphery
register and the alive notification from the firmware.
We have a check in iwl_pcie_tx_start that ensures that
they are the same.
When we resume from WoWLAN, the firmware may have crashed
for whatever reason. In that case, the whole device may be
reset which means that the periphery register will hold a
meaningless value. When we come to compare
trans_pcie->scd_base_addr (which really holds the value we
had when we loaded the WoWLAN firmware upon suspend) and
the current value of the register, we don't see a match
unsurprisingly.
Trick the check to avoid a loud yet harmless WARN.
Note that when the WoWLAN has crashed, we will see that
in iwl_trans_pcie_d3_resume which will let the op_mode
know. Once the op_mode is informed that the WowLAN firmware
has crashed, it can't do much besides resetting the whole
device.

BUG=chrome-os-partner:34821
TEST=suspend/resume with WoWLAN, check that the pcie/tx.c
WARN_ON() in case the device somehow got reset is gone

Change-Id: Id0c83d00e73eb9b7e665e39daa799db26dc715eb
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/244812
Tested-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Johannes Berg <johannes.berg@intel.com>