CHROMIUM: mtd: spi-nand: Return number of corrected bitflips when reading
This commit fixes a bug in the current SPI NAND implementation, which
misses the update of the 'ecc_stats.corrected' field in the MTD main structure.

This field holds the number of corrected ECC bitflips, as reported by
the device's internal ECC controller.

Without this patch, the bitflips aren't reported to upper layers such
as UBI. UBI uses the number of corrected bitflips to decide if an
eraseblock needs scrubbing.

BUG=chrome-os-partner:35348
TEST=nandtest on Pistachio BuB

Change-Id: I772b51be9c42115fe31f3a1f9ba93b70783fa2db
Signed-off-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/245314
Reviewed-by: Daniel Ehrenberg <dehrenberg@chromium.org>
Tested-by: Ionela Voinescu <ionela.voinescu@imgtec.com>
Commit-Queue: James Hartley <james.hartley@imgtec.com>