UPSTREAM: rhashtable: Fix potential crash on destroy in rhashtable_shrink
The current being_destroyed check in rhashtable_expand is not
enough since if we start a shrinking process after freeing all
elements in the table that's also going to crash.

This patch adds a being_destroyed check to the deferred worker
thread so that we bail out as soon as we take the lock.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 28134a53d624ae7e90fff8500b25b3add4d40b92)

BUG=chromium:556861
TEST=builds all four kernel flavors based on chromeos-3.18 branch

Change-Id: I1c77852a345572961101171bf344d3006515a8f0
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/314162
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>