CHROMIUM: ARM: dts: rockchip: remove hpd pin from speedy-rev1
The 2 kinds of panel cable of speedy-rev1 have some problem, one is disconnect,
and the other one is short, so the hpd signal are both incorrect. Remove the
references of hpd from speedy-rev1 dts, that will keep the hpd pin for a gpio
function and make panel display normally.

BUG=chrome-os-partner:36331
TEST=Build and boot, the speedy-rev1 can display normally.

Change-Id: I548eb3bfbd77aad94e6439ad2e40e36aad80bdfa
Signed-off-by: Chris Zhong <zyw@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/246090
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Olof Johansson <olofj@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>