CHROMIUM: iwl7000: call ieee80211_scan_completed() even if scan abort fails
A scan abort command failure is not that unusual, since we may try to
send it after the scan has actually completed but before we received
the completed notification from the firmware.  The scan abort can also
fail for other reasons, such as a timeout.  In such cases, we should
clear things up so the next scans will work again.  To do so, don't
return immediately in case of failures, but call
ieee80211_scan_completed() and clear the scan_status flags.

BUG=chrome-os-partner:36253
TEST=cause scan offload abort to fail (hack) and check that scan
     doesn't get stuck completely

Change-Id: I46886620f1ad28abbb4b04efb4ea007c1413bbf6
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/246485
Tested-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Johannes Berg <johannes.berg@intel.com>