CHROMIUM: mips: dts: Enable the in-flash bad block table for SPI NAND
Using an in-flash bad block table (BBT) has a few advantages. First,
it allows to reduce the boot time. Without an in-flash BBT, the kernel
needs to scan the whole device to construct the in-memory BBT.

Also, it's not certain that bad blocks will be able to be marked
as bad in the OOB (because they are bad!). This doesn't seem to be
an issue for SPI NAND, as bad blocks are read as filled with zeroes.

By using the 'nand-on-flash-bbt' devicetree property, the SPI NAND driver
will enable the in-flash BBT. When a block is found to be bad it will
be marked in the BBT, and also in the block's first page OOB.

This bad block marker in the OOB is a requirement for depthcharge to
detect the bad block and skip it.

BUG=chrome-os-partner:35348
TEST=Mark some blocks as bad from Linux (using custom tool) and
verified they are detected by depthcharge.

Change-Id: I933097e7d3396f40e6764b5ab7249610c673ed70
Signed-off-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/249380
Reviewed-by: Daniel Ehrenberg <dehrenberg@chromium.org>
Tested-by: Daniel Ehrenberg <dehrenberg@chromium.org>