UPSTREAM: timekeeping: Pass readout base to update_fast_timekeeper()
Modify update_fast_timekeeper() to take a struct tk_read_base
pointer as its argument (instead of a struct timekeeper pointer)
and update its kerneldoc comment to reflect that.

That will allow a struct tk_read_base that is not part of a
struct timekeeper to be passed to it in the next patch.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Acked-by: John Stultz <john.stultz@linaro.org>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/rafael/linux-pm
br: linux-next
(cherry picked from commit affe3e85ae78507cc953f3f700e0644e50844cff)

BUG=None
TEST=1) Change power state of system to "freeze"
     2) Use echo freeze > /sys/power/state
     3) Check whether system wakes up on any interrupt(set rtc timer)

Change-Id: I7be799b22622fc51281f375812a6b5e44ec4ba50
Signed-off-by: Jay Patel <jay.p.patel@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/282853
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Prathyushi Nangia <prathyushi.nangia@intel.com>