CHROMIUM: rockchip: dmc: Turn off mmu for rate change
This turns off the mmu for changing the dmc clk rate. It also separates out the
accesses to dram in the dmc_set_rate_in_sram function to avoid an expensive
cache flush, which we don't have the time for. This is because we need to
complete before vblank is finished.

BUG=chrome-os-partner:31440
TEST=boot jerry

Change-Id: Idbb8d14c7762e252e9fae04ec9299cfa836346ca
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/248490
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>