UPSTREAM: thermal: Introduce dummy functions when thermal is not defined
When CONFIG_THERMAL is not enabled, it is better to introduce
equivalent dummy functions in the exported header than to
introduce #ifdeffery in drivers using the function.

This will prevent issues such as that reported in:
http://www.spinics.net/lists/linux-next/msg31573.html

While at it switch over to IS_ENABLED for thermal macros
to allow for thermal framework to be built as framework
and relevant APIs be usable by relevant drivers as a result.

Reported-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Eduardo Valentin <edubezval@gmail.com>
(cherry picked from commit 12ca7188468ee29c4e717f73db4bf43c90954fc7)
Signed-off-by: Kevin Hilman <khilman@linaro.org>

BUG=chrome-os-partner:42804
TEST=`emerge-oak chromeos-kernel-3_18`

Change-Id: I9474a89f090363eaf1ba5f347600fbf54ac05da0
Signed-off-by: Dawei.Chien <dawei.chien@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/283820
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Commit-Queue: Wei-Ning Huang <wnhuang@chromium.org>