CHROMIUM: vb2: don't adjust bytesused if 0 is passed from userspace.
Mainline vb2 adjusts bytesused to be equal to the total length of a
buffer/plane if the userspace passes bytesused=0. This is to correct
the behavior of erroneous userspace.

However, this also breaks some drivers (notably s5p-mfc and go2001) and
current Chrome, which use the old way to flush the codec by queuing
an empty bitstream buffer.

Revert this behavior for now, until this gets properly handled in mainline.

Signed-off-by: Pawel Osciak <posciak@chromium.org>

BUG=None
TEST=vdatest on x86 and ARM

Change-Id: I689dab1811022cde8620b7950e8353c0974511b4
Reviewed-on: https://chromium-review.googlesource.com/250760
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>
Trybot-Ready: Pawel Osciak <posciak@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>