drm/tegra: dc: Reset state's active_changed field
Commit eab3bbeffd15 ("drm/atomic: Add drm_crtc_state->active") added the
field to track the DPMS state. However, the Tegra driver was in modified
in parallel and subclasses the CRTC atomic state, so needed to duplicate
the code in the atomic helpers. After the addition of the active_changed
field it became out of sync and doesn't reset it when duplicating state.

This causes a full modeset on things like page-flips, which will in turn
cause warnings due to the VBLANK machinery being disabled when it really
should remain on.

Tested-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit 567a3cd1d75e3f282759af3bd05a7369daef2b2a)
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>