FROMLIST: mmc: dw_mmc: Make sure we only adjust the clock when power is on
It appears that we can confuse things if we try to turn on the MMC
clock when the power is off.  Adjust is so that we turn the clock on
(using dw_mci_setup_bus) after power is all the way on and we turn the
clock off before the power goes off.

Conflicts:
    drivers/mmc/host/dw_mmc.c
...due to tuning stuff in our local tree; trivial to resolve.

BUG=chrome-os-partner:35437
TEST=Manual:
  cd /sys/bus/platform/drivers/dwmmc_rockchip
  for i in $(seq 1 10000); do
    echo "========================" $i
    echo ff0c0000.dwmmc > unbind
    sleep .5
    echo ff0c0000.dwmmc > bind
    sleep 2
  done

Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/5858261/)
Change-Id: I0c8be4a735b1bd2bf8516ea28dfef0a4bf88785c
Reviewed-on: https://chromium-review.googlesource.com/251341
Reviewed-by: Olof Johansson <olofj@chromium.org>