FROMLIST: mmc: dw_mmc: Give a good reset after we give power
We should give dw_mmc a good reset after we apply power.  On some
boards vqmmc may actually be connected to the IP block in the SoC so
it's good to reset after power comes in.

Without this we sometimes see failures enumerating cards on rk3288.

BUG=chrome-os-partner:35437
TEST=Manual:
  cd /sys/bus/platform/drivers/dwmmc_rockchip
  for i in $(seq 1 10000); do
    echo "========================" $i
    echo ff0c0000.dwmmc > unbind
    sleep .5
    echo ff0c0000.dwmmc > bind
    sleep 2
  done

Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/5858281/)
Change-Id: I45314ae4b7bf679454035c2fd18b37a6a343f393
Reviewed-on: https://chromium-review.googlesource.com/251344
Reviewed-by: Olof Johansson <olofj@chromium.org>