CHROMIUM: mmc: dw_mmc: Only change phase when clock changes
On the SDIO slot on rk3288 boards we keep power across suspend/resume.
That means that we don't go through tuning again, but set_ios is
called.  That was clobbering the tuning that we did and resetting us
to the default phase.  This could possibly cause problems with WiFi
after suspend/resume.

Let's fix that by only resetting to the default phase when the clock
changes.

BUG=chrome-os-partner:36961
TEST=suspend/resume, grep sdio0_sample /sys/kernel/debug/clk/clk_summary

Change-Id: Idf0d469296c3bb6f4ed269ed73c8270e7276a57d
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/252091
Reviewed-by: Alexandru Stan <amstan@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>