CHROMIUM: iio: adc: Fix incorrect use of ADC Powerup/down register
To match with the register definitions in the device spec, change the
register name from CC10001_ADC_POWER_UP to CC10001_ADC_POWER_DOWN and
also change the names of register fields as per the regiser name.

At present we are incorrectly setting the register to 0x1 to Powerup
the ADC. Since it is a active high PowerDown register, we need to set
the register to 0x0 to actually Powerup. Writing value 0x1 to the
register actually enables Powerdown mode.

BUG=chrome-os-partner:35348
TEST=Tested on Pistachio BuB

Change-Id: I050fe18c6ac739b3ffd1ebe29a5770a48b835aa2
Signed-off-by: Naidu Tellapati <naidu.tellapati@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/252610
Reviewed-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: James Hartley <james.hartley@imgtec.com>