CHROMIUM: iio: adc: Fix the issue related to channel number
Fix the issue related to channel number and scan index in IIO buffer
capture mode.

When we enable all the channels from the DTS, both the scan index
and corresponding channel number are the same. But the issue here is
that when we disable some of the channels from the DTS, the scan index
and corresponding channel number are not the same. The IIO buffer
trigger handler can give us only scan index. Store the channel number
inside the device initialization function corresponding to each scan
index in an array inside device structure and later use it for
retrieving channel number from scan index inside the trigger handler
for starting conversion.

BUG=chrome-os-partner:35348
TEST=Tested on Pistachio BuB

Change-Id: I99d382c2827c03c54a29fbb7f12d1424dda5bd2b
Signed-off-by: Naidu Tellapati <naidu.tellapati@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/252630
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: James Hartley <james.hartley@imgtec.com>