UPSTREAM: ALSA: hda - Allocate hda_pcm objects dynamically
So far, the hda_codec object kept the hda_pcm list in an array, and
the codec driver was expected to assign the array.  However, this
makes the object life cycle management harder, because the assigned
array is freed at the codec driver detach while it might be still
accessed by the opened streams.

In this patch, we allocate each hda_pcm object dynamically and manage
it as a linked list.  Each object has a kref refcount, and both the
codec driver binder and the PCM open/close touches it, so that the
object won't be freed while in use.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git
br: master
(cherry picked from commit bbbc7e8502c95237dbd86cc4d0a12ca9a6b18c8f)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=None
TEST=with the whole series applied, the audio subsystem is stable.
TEST=there are no regressions observed on BSW.

Change-Id: Ia3cf88fd7ce93790f3048c76699243ac542a2b9f