FROMLIST: regulator: core: Fix enable GPIO reference counting
It is possible for _regulator_do_enable() to be called for an
already-enabled rdev, like in regulator_suspend_finish().  If we were
using an enable pin (rdev->ena_pin is set) then we'd end up
incrementing the reference count in regulator_ena_gpio_ctrl() over and
over again without a decrement.  That prevented the GPIO from going to
the "off" state even after all users were disabled.

Fix this by avoiding the call to regulator_ena_gpio_ctrl() when it's
not needed.

BUG=chrome-os-partner:37187
TEST=suspend_stress_test -c1 && echo 1 > /sys/kernel/debug/mwifiex/mlan0/reset

Signed-off-by: Doug Anderson <dianders@chromium.org>
Fixes: 967cfb18c0e3 ("regulator: core: manage enable GPIO list")
(am from https://patchwork.kernel.org/patch/5903071/)
Change-Id: I09aa847dca5cba73f9b865491b45a4f4621d5c20
Reviewed-on: https://chromium-review.googlesource.com/254661
Reviewed-by: Paul Stewart <pstew@chromium.org>