UPSTREAM: cpuidle / sleep: Use broadcast timer for states that stop local timer
Commit 381063133246 (PM / sleep: Re-implement suspend-to-idle handling)
overlooked the fact that entering some sufficiently deep idle states
by CPUs may cause their local timers to stop and in those cases it
is necessary to switch over to a broadcast timer prior to entering
the idle state.  If the cpuidle driver in use does not provide
the new ->enter_freeze callback for any of the idle states, that
problem affects suspend-to-idle too, but it is not taken into account
after the changes made by commit 381063133246.

Fix that by changing the definition of cpuidle_enter_freeze() and
re-arranging of the code in cpuidle_idle_call(), so the former does
not call cpuidle_enter() any more and the fallback case is handled
by cpuidle_idle_call() directly.

Fixes: 381063133246 (PM / sleep: Re-implement suspend-to-idle handling)
Reported-and-tested-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/rafael/linux-pm
br: linux-next
(cherry picked from commit ef2b22ac540c018bd574d1846ab95b9bfcf38702)

BUG=None
TEST=1) Change power state of system to "freeze"
     2) Use echo freeze > /sys/power/state
     3) Check whether system wakes up on any interrupt(set rtc timer)

Change-Id: I848d7cc7b77638d581f9a98f7e1f456ce0333660
Signed-off-by: Jay Patel <jay.p.patel@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/282861
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Prathyushi Nangia <prathyushi.nangia@intel.com>