CHROMIUM: drm: Shrink vblank off delay to 50ms
The patch to reduce the vblank off delay got lost in the 3.8 -> 3.10
transition. But for 3.14 if we want to use PSR efficiently we should
actually reduce it all the way to 50ms. Since it saves quite a bit of
power, let's put it back and use 50ms instead of 1s.

i915 seems to work, hopefull this doesn't cause too much trouble with
other drivers.

BUG=none
TEST=builds and runs on samus

Change-Id: I74a35e4b66ed10eab259d3316dc45b93ae8abaa3
Reviewed-on: https://chromium-review.googlesource.com/206533
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>