BACKPORT: drm/nouveau/platform: fix probe error path
A "return 0" found its way in the middle of the error path of
nouveau_platform_probe(), remove it as it will make the kernel crash if
we try to unload the module afterwards.

While we are at it, also remove the IOMMU domain if it has been created,
as we should.

Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>
Signed-off-by: Ben Skeggs <bskeggs@redhat.com>
(cherry picked from commit bbf2e92f914019324bb4351fcf0db9f41329a7c3)
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>

Change-Id: Ic6c95bb9337f8fb3cf11ccf0ce3b3fed2ef5a8b7
Reviewed-on: https://chromium-review.googlesource.com/269491
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>