CHROMIUM: iwl7000: add new TLV capability flag for gscan support
Gscan is a scan feature which is supported on certain devices only,
hence the need for a TLV flag for it. For devices that support gscan
store the gscan capabilities advertised by the FW so the driver can
report it to upper layers.

Signed-off-by: Avraham Stern <avraham.stern@intel.com>
iwl7000-tree: 82a3ac68ac1e4d8e56df48d50f50ffcb61c2a2c4