UPSTREAM: rhashtable: Allow hashfn to be unset
Since every current rhashtable user uses jhash as their hash
function, the fact that jhash is an inline function causes each
user to generate a copy of its code.

This function provides a solution to this problem by allowing
hashfn to be unset.  In which case rhashtable will automatically
set it to jhash.  Furthermore, if the key length is a multiple
of 4, we will switch over to jhash2.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Acked-by: Thomas Graf <tgraf@suug.ch>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 31ccde2dacea8375c3a7d6fffbf0060ee0d40214)

BUG=chromium:556861
TEST=builds all four kernel flavors based on chromeos-3.18 branch

Change-Id: I5c7c6446a1cc0786e366317fc5ac24a9b40a9ec4
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/314211
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>