CHROMIUM: ARM64: dts: tegra210: Add "nvidia,tegra132-sdhci" to sdhci nodes
Since the controllers on Tegra210 seem to be compatible with Tegra132,
at least in terms of basic operation, there is no real need to push new
compatible string to the kernel until we find that it is needed for some
reason. This patch is adding "nvidia,tegra132-sdhci" to the list of
compatible strings of sdhci nodes of Tegra210 to let the unmodified
kernel driver handle the controllers.

BUG=none
TEST=booted foster from sdhci

Change-Id: I8688f87bd345c1d7b6caed32eb5a025823bb7fc9
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chrome-internal-review.googlesource.com/210057
Reviewed-by: Tomasz Figa <tfiga@google.com>
Commit-Queue: Tomasz Figa <tfiga@google.com>
Tested-by: Tomasz Figa <tfiga@google.com>
Reviewed-by: David Riley <davidriley@google.com>
Reviewed-on: https://chromium-review.googlesource.com/267481
Tested-by: David Riley <davidriley@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>
Commit-Queue: David Riley <davidriley@chromium.org>