arm: bl_switcher: Kill tick suspend hackery
Use the new tick_suspend/resume_local() and get rid of the homebrewn
implementation of these in the ARM bL switcher.  The check for the
cpumask is completely pointless.  There is no harm to suspend a per
cpu tick device unconditionally.  If that's a real issue then we fix
it proper at the core level and not with some completely undocumented
hacks in some random core code.

Move the tick internals to the core code, now that this nuisance is
gone.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Cc: Russell King <rmk+kernel@arm.linux.org.uk>
[ rjw: Rebase, changelog ]
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Nicolas Pitre <nicolas.pitre@linaro.org>