CHROMIUM: cros-ec: Force flash write if in RO but RW hash matches.
The signature is stored at the tail of the RW flash.  If a previous
flash of RW was successful in updating but subsequently the signature
is erased the device can end up in a state where its stuck in RO but
with RW hash match.  This state is persistent across power cycles and
therefore leaves the device permanently running RO code.

CL changes decision logic to include this scenario and forces re-flash
in that case.

Signed-off-by: Todd Broch <tbroch@chromium.org>

BUG=chrome-os-partner:33947
TEST=manual, on zinger that has been successfully updated to latest FW
erase its signature by sending the VDO_CMD_ERASE_SIG VDM and see it
successfully re-flash.

Change-Id: I853eee973b6ed6f8c93f57873f44c04fcddf5e76
Reviewed-on: https://chromium-review.googlesource.com/263064
Tested-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Shawn N <shawnn@chromium.org>
Commit-Queue: Todd Broch <tbroch@chromium.org>
Reviewed-by: Alec Berg <alecaberg@chromium.org>