FROMLIST: ACPI / PM: Attach ACPI power domain only once
Some devices, like MFD subdevices, share a single ACPI companion device so
that they are able to access their resources and children. However,
currently all these subdevices are attached to the ACPI power domain and
this might cause that the power methods for the companion device get called
more than once.

In order to solve this we attach the ACPI power domain only to the first
physical device that is bound to the ACPI companion device. In case of MFD
devices, this is the parent MFD device itself.

Acked-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Link: https://patchwork.kernel.org/patch/6607681/
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

Conflicts:
	(Conflict due to minor difference in code. Removing ret because
	it is just an unused variable at this point.)
	drivers/acpi/scan.c

BUG=chrome-os-partner:40857
TEST=In the output of lspci -v, intel-lpss name gets attached to most
   "Signal processing controller" sections.
TEST=Two new lines "Synopsys Designware I2C adapter" appear in:
 - cat /sys/bus/i2c/devices/*/name
 - and in the output "i2cdetect -l" (i2cdetect on "test" image at least)
TEST=Two new "i2c_designware" lines appear in /proc/interrupts

Change-Id: I59bf2babafe6217f0c261c644f4e36940b3fc7ae
Reviewed-on: https://chromium-review.googlesource.com/282872
Reviewed-by: Joe Konno <joe.konno@linux.intel.com>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Tested-by: Philip Hanson <philip.hanson@intel.com>
Commit-Queue: Vedang Patel <vedang.patel@intel.com>