FROMLIST: clk: tegra: pll: Add logic for out-of-table rates for T210
For Tegra210, the logic to calculate out-of-table rates is different
from previous generations. Add callbacks that can be overridden to
allow for different ways of calculating rates. Default to
_cal_rate when not specified.

This patch also includes a new flag which is used to set which method
of fixed_mdiv calculation is used. The new method for calculating the
fixed divider value for M can be more accurate especially when
fractional dividers are in play. This allows for older chipsets to use
the existing logic and new generations to use a newer version which
may work better for them.

Based on original work by Aleksandr Frid <afrid@nvidia.com>

(git am from https://patchwork.ozlabs.org/patch/471421/)

BUG=chrome-os-partner:39779
TEST=build kernel and boot-test to verify clocks.

Change-Id: If5b8c1477e1aabbc2377df914c948c5353e08486
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/268329
Reviewed-by: Benson Leung <bleung@chromium.org>
Tested-by: Bill Huang <bilhuang@nvidia.com>
Commit-Queue: Bill Huang <bilhuang@nvidia.com>