CHROMIUM: cros_usbpd-charger: Cached power supply property
When insert / removing zinger, there are several query to ec in
burst for about 100-200 milliseconds. This is because userland
update several propertys when power supply status changed and
kernel driver asks ec each time we receive the inquiry.

Instead of asking ec every time when userland queries the
charger property. This patch return cached result if it is
within 500 milliseconds of last update.

BUG=None
TEST=Plug zinger. EC does not get spammed.

Change-Id: I7eef0aed359fa75896d4d8a147c063de1b05af4f
Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/263620
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>