CHROMIUM: iwl7000: idi: enable DBGM
Enable DBGM over IDI (as rx frames, using out of band channel)

There are 2 debug buffers. Each one has associated event,
that when invoked will configure the kick2 (through LL)
to pass the relevant debug buffer to the rx fifo.

In order to use the debug buffers, we have to decrease
the TBs count. In order to minimize the changes to the
IDI layer, redefine some macros (which used to have
a const value) to point to a "layout struct", that
can be changed in runtime through the ini file.

Signed-off-by: Eliad Peller <eliadx.peller@intel.com>
Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
Signed-off-by: Idan Kahlon <idanx.kahlon@intel.com>
iwl7000-tree: 5a4a50e99a33ad4816e6bbf80a3e4ebac13ba462