UPSTREAM: arm64: psci: kill psci_power_state
A PSCI 1.0 implementation may choose to use the new extended StateID
format, the presence of which may be queried via the PSCI_FEATURES call.
The layout of this new StateID format is incompatible with the existing
format, and so to handle both we must abstract attempts to parse the
fields.

In preparation for PSCI 1.0 support, this patch introduces
psci_power_state_loses_context and psci_power_state_is_valid functions
to query information from a PSCI power state, which is no longer
decomposed (and hence the pack/unpack functions are removed). As it is
no longer decomposed, it is now passed round as an opaque u32 token.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Acked-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Tested-by: Hanjun Guo <hanjun.guo@linaro.org>
Cc: Will Deacon <will.deacon@arm.com>
(cherry picked from commit c8cc42737788537ebef810ee22400f757e1819ca)
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:41584
TEST=build and boot oak-rev2 & smaug w/ PSCI-v1.0 enabled FW image
TEST=powerd_dbus_suspend ; resume using keyboard and touchpad

Change-Id: I17c57028ee3392ed937d1a5d692ac755f08084a5
Reviewed-on: https://chromium-review.googlesource.com/280471
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Joseph Lo <josephl@nvidia.com>
Tested-by: Joseph Lo <josephl@nvidia.com>