UPSTREAM: PM / sleep: Refine diagnostic messages in enter_state()
Some of the system suspend diagnostic messages related to
suspend-to-idle refer to it as "freeze sleep" or "freeze state"
while the others say "suspend-to-idle".  To reduce the possible
confusion that may result from that, refine the former either to
say "suspend to idle" too or to make it clearer that what is printed
is a state string written to /sys/power/state ("mem", "standby",
or "freeze").

BUG=chrome-os-partner:43079
TEST=None

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit a921504588cfe0e11d9306536bdf131b5b957fd3)
Signed-off-by: Derek Basehore <dbasehore@chromium.org>

Change-Id: I66d2dd11627a781fde4d846a6af1c7d9d3abcc64
Reviewed-on: https://chromium-review.googlesource.com/308177
Commit-Ready: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>