UPSTREAM: x86/platform/atom/punit: Add Punit device state debug driver
The patch adds a debug driver, which dumps the power states
of all the North complex (NC) devices. This debug interface is
useful to figure out the devices,  which blocks the S0ix
transitions on the platform. This is extremely useful during
enabling PM on customer platforms and derivatives.

This submission is based on the submission from Mahesh Kumar P:

  https://lkml.org/lkml/2014/11/5/367

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Cc: Borislav Petkov <bp@alien8.de>
Cc: H. Peter Anvin <hpa@zytor.com>
Cc: Mahesh Kumar P <mahesh.kumar.p@intel.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: pebolle@tiscali.nl
Link: http://lkml.kernel.org/r/1430939754-6900-2-git-send-email-srinivas.pandruvada@linux.intel.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>

BUG=chrome-os-partner:42702
TEST=cat /sys/kernel/debug/punit_atom/dev_power_state

(cherry picked from commit 956079e081427fe0c929eb284ab7e39f9b8e2023)
Change-Id: Ic4b26bbe7fe4abbc7d31b4b06ca5ef9d36d37419
Signed-off-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/298131
Tested-by: Eduard G Florinescu <eduard.g.florinescu@intel.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>