UPSTREAM: mmc: core: Enable / disable re-tuning
Enable re-tuning when tuning is executed and
disable re-tuning when card is no longer initialized.

In the case of SDIO suspend, the card can keep power.
In that case, re-tuning need not be disabled, but, if
a re-tuning timer is being used, ensure it is disabled
and assume that re-tuning will be needed upon resume
since it is not known how long the suspend will last.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
(cherry picked from commit 79d5a65aeea43920bf3ff60791f317570dd6f54f
 https://git.linaro.org/people/ulf.hansson/mmc.git next)

BUG=chrome-os-partner:41746
TEST=Set mmc0/mmc1 max-frequency = <200000000> in mt8173-oak.dtsi

Change-Id: I14064d8a6e6699f20d06012ea381bc134ef24e81
Signed-off-by: Chaotian Jing <chaotian.jing@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/306513
Commit-Ready: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>