CHROMIUM: arm64: dts: mt8173-oak: Add a cros-ec-spi i2c_tunnel node
Add a cros-ec-spi tunnel node on the "spi_gpio" bus and corresponding
pinctrl settings.

The cros-ec-spi tunnel allows the AP to communicate with the cros-ec using
i2c commands over a SPI bus.

The MT8173 has a native spi port, but the plan is to use it for an audio
codec.  Thus, we attach the cros-ec-spi tunnel on a bit-banged (GPIO) SPI.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:38746
TEST=build and boot oak-rev0 & -rev1

Change-Id: Id48ce1092a7de896708d4974b804d84a40ea39ba
Reviewed-on: https://chromium-review.googlesource.com/270731
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Trybot-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>