CHROMIUM: regulator: max8973: add external pin gpio
Both the max77621 and max8973 have pins that can externally
control the state of regulator.  For max77621, there is the
nSHDN pin.  For max8973, it is the EN pin.  However, it may
be desired that these pins not influence the state of the
regulator and the regulator be enabled/disabled only through
the I2C registers.

This allows these pins to be configured to a non-forcing
state if they can be supplied to the driver through a
configurable gpio.  It will also only be done if
"maxim,externally-enable" is not supplied.  In the case of
max77621, the external pin gpio should control the nSHDN pin
and will be driven high.  In the case of the max8973, the
external pin gpio should control the EN pin and be driven low.

BUG=chrome-os-partner:40325
TEST=Boot to console

Change-Id: Ibb42b486223ee51861b31b287df6a703091c6797
Signed-off-by: Christopher Freeman <cfreeman@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/271440
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>