CHROMIUM: arm64: dragon_defconfig: Disable TEGRA_HOST1X_FIREWALL
The host1x firewall scans through all of the command buffers and finds
writes towards address registers.  In case one is found, we check that the
register gets written only through the relocation mechanism.  This was
introduced when there was no SMMU support, to prevent applications from
stealing data.

With SMMU, we don't need this.  Applications will not be able to use
engines for fetching data from arbitrary memory locations.  In addition,
this mechanism isn't implemented for certain tegradrm clients like the
ones supported in flcn.c.

BUG=chrome-os-partner:39847
TEST=Built and booted on smaug

Change-Id: I33188179266e64f47c81b299859f6cdbd89c06a1
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/269896
Reviewed-by: Viranjan Pagar <vpagar@nvidia.com>
Tested-by: Viranjan Pagar <vpagar@nvidia.com>
Reviewed-by: Benson Leung <bleung@chromium.org>