UPSTREAM: cfg80211: properly send NL80211_ATTR_DISCONNECTED_BY_AP in disconnect
commit 80279fb7ba5b71981a60988b0307afa43f78f6b1 upstream

When we disconnect from the AP, drivers call cfg80211_disconnect().
This doesn't know whether the disconnection was initiated locally
or by the AP though, which can cause problems with the supplicant,
for example with WPS. This issue obviously doesn't show up with any
mac80211 based driver since mac80211 doesn't call this function.

Fix this by requiring drivers to indicate whether the disconnect is
locally generated or not. I've tried to update the drivers, but may
not have gotten the values correct, and some drivers may currently
not be able to report correct values. In case of doubt I left it at
false, which is the current behaviour.

For libertas, make adjustments as indicated by Dan Williams.

Reported-by: Matthieu Mauger <matthieux.mauger@intel.com>
Tested-by: Matthieu Mauger <matthieux.mauger@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>

BUG=chrome-os-partner:48225
TEST=autotest using brcmfmac with 4371 device

Change-Id: Ifcada743fd40a86972c8bfa45f1b1169ebe35ac9
[arend@broadcom.com: also apply change to bcmdhd]
Signed-off-by: Arend van Spriel <arend@broadcom.com>
Reviewed-on: https://chromium-review.googlesource.com/320990
Commit-Ready: Arend van Spriel <aspriel@gmail.com>
Tested-by: Arend van Spriel <aspriel@gmail.com>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>