FROMLIST: mfd: make mfd_remove_devices() iterate in reverse order
The newly introduced device_for_each_child_reverse() would be used when MFD
core removes the device.

After this patch applied the devices will be removed in a reversed order. This
behaviour is useful when devices have implicit dependency on order, i.e.
consider MFD device with serial bus controller, such as SPI, and DMA IP that is
attached to serial bus controller: before remove the DMA driver we have to be
ensured that no DMA transfers is ongoing and the requested channel are unused.

Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Link: https://patchwork.kernel.org/patch/6607751/
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=chrome-os-partner:40857
TEST=In the output of lspci -v, intel-lpss name gets attached to most
   "Signal processing controller" sections.
TEST=Two new lines "Synopsys Designware I2C adapter" appear in:
 - cat /sys/bus/i2c/devices/*/name
 - and in the output "i2cdetect -l" (i2cdetect on "test" image at least)
TEST=Two new "i2c_designware" lines appear in /proc/interrupts

Change-Id: Idc2085c6a508d2773a8376bd7d2f200f3963578d
Reviewed-on: https://chromium-review.googlesource.com/282876
Reviewed-by: Joe Konno <joe.konno@linux.intel.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Philip Hanson <philip.hanson@intel.com>
Commit-Queue: Vedang Patel <vedang.patel@intel.com>