CHROMIUM: config: enable CONFIG_LOCALVERSION_AUTO
Enabled CONFIG_LOCALVERSION_AUTO so that the kernel revision includes
the kernel git hash. Note that if the kernel has been cros_workon'ed
then a .scmversion file be will dropped in the output directory to
exclude the git sha from kernel version (this is useful for
developers doing local kernel builds).

BUG=chrome-os-partner:40556
TEST=Couple of test cases:
1. for cros_workon kernel build, verify that the kernel version
is of the form 3.18.0 and doesn't include git hash in the
kernel version.
2. for non-cros_workon kernel builds the kernel version should include
the git sha.

Change-Id: Ib0e9111163e25dec2a37953cc74f6b3526078f2f
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/273278
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>

Conflicts:
	chromeos/config/base.config
[rebase44(filbranden): Removed changes to the file from this commit.]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>