FROMLIST: bcb: Android Bootloader Control Block driver
Android userspace tells the kernel that it wants to boot into recovery
or some other non-default OS environment by passing a string argument
to reboot(). It is left to the OEM to hook this up to their specific
bootloader.

This driver uses the bootloader control block (BCB) in the 'misc'
partition, which is the AOSP mechanism used to communicate with
the bootloader. Writing 'bootonce-NNN' to the command field
will cause the bootloader to do a oneshot boot into an alternate
boot label NNN if it exists. The device and partition number are
passed in via kernel command line.

It is the bootloader's job to guard against this area being uninitialzed
or containing an invalid boot label, and just boot normally if that
is the case.

Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>

BUG=none
TEST=adb reboot recovery

Change-Id: I49faf348bef7817f9fe1a6c2691acda01bee7bac
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/275278
Reviewed-by: Sameer Nanda <snanda@chromium.org>