UPSTREAM: selinux: enable per-file labeling for debugfs files.
Add support for per-file labeling of debugfs files so that
we can distinguish them in policy.  This is particularly
important in Android where certain debugfs files have to be writable
by apps and therefore the debugfs directory tree can be read and
searched by all.

Since debugfs is entirely kernel-generated, the directory tree is
immutable by userspace, and the inodes are pinned in memory, we can
simply use the same approach as with proc and label the inodes from
policy based on pathname from the root of the debugfs filesystem.
Generalize the existing labeling support used for proc and reuse it
for debugfs too.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>
Signed-off-by: Paul Moore <pmoore@redhat.com>
(cherry picked from commit 134509d54e4e98888be2697a92cb4b48957b792b)
Signed-off-by: Stephen Barber <smbarber@chromium.org>

BUG=none
TEST=kernel compiles and boots

Change-Id: Ib605028e3f43be25724a3e8ec60d40f54e52e7ac
Reviewed-on: https://chromium-review.googlesource.com/281743
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Stephen Barber <smbarber@chromium.org>