UPSTREAM: ALSA: hda - Don't actually write registers for caps overwrites
Along with the transition to regmap for managing the cached parameter
reads, the caps overwrite was also moved to regmap cache.  The cache
change itself works, but it still tries to write the non-existing verb
(the HDA parameter is read-only) wrongly.  It's harmless in most
cases, but some chips are picky and may result in the codec
communication stall.

This patch avoids it just by adding the missing flag check in
reg_write ops.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/tiwai/sound.git
br: for-next
(cherry picked from commit 98a226ed21949601b270f7ea20abc9f72f7b0be9)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=None
TEST=with the whole series applied, the audio subsystem is stable.
TEST=there are no regressions observed on BSW.

Change-Id: I4c1e3246859184873bcc01f1fdbd1a503ef633fa