CHROMIUM:Revert "config: enable CONFIG_PPP options"
This reverts commit 6634f8d246beac5693b51002c40ca1f2f6d828f7.
When CL:272741 by Eric Caruso was submitted, kernel_ConfigVerifyPPP
would pass. But the test is not part of bvt-inline anymore, and is
part of experimental. It is only supposed to pass when USE=ppp is set,
and hence the test is put in "experimental". Hence kernel_ConfigVerify
and security_ProtocolFamilies, which are partof bvt-inline, will pass,
when PPP is disabled by default.

BUG=chrome-os-partner:41202
TEST=test_that -b cyan IP security_ProtocolFamilies kernel_ConfigVerify

Change-Id: Ied4d231a70d53dd233a6845f8ea381abc7063704
Signed-off-by: Arindam Roy <arindam.roy@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/276681
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>