HACK: drm/nouveau/clk: disable two highest pstates for GM20B
On Smaug the GPU rail is 16A and can't support the frequency that is higher
than 844.8MHz for GM20B due to EDP design. The frequency should be capped
by GPU EDP driver which is not yet ready. So just remove the highest two
pstates for now.

BUG=chrome-os-partner:41434
TEST=Enable GPU DVFS and UI is stable

Change-Id: I3f4f7de3b9fbac45f8b92ed3d27fe2eb30c053ae
Signed-off-by: Vince Hsu <vinceh@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/277628
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>