drm/nouveau/volt: take thermal into account for voltage calculation
The voltage level for a thermal range should be the higher one of the
calculated results of the lower and upper bounds. For now we only consider
the lowest thermal range which needs the highest voltage. Later we will
support all the thermal ranges and GPU thermal throttling.

BUG=chrome-os-partner:41434
TEST=Enable GPU DVFS and UI is stable

Change-Id: If41c861379d392a1144606d3fa9769adeb4c97db
Signed-off-by: Vince Hsu <vinceh@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/277626
Tested-by: Vince Hsu <vince.h@nvidia.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>