CHROMIUM: arm64: call idle notifiers when CPU goes idle
The interactive governor's timer task keeps tuning the CPU frequency
up until it hits the highest frequency, at which time the timer task
stops itself.  The timer task is re-enabled after the CPU goes idle,
which is when the target frequency for the CPU gets re-evaluated.

We should call the idle notifiers on arm64 when CPU goes idle so the
interactive governor's timer task can be resurrected.

BUG=chrome-os-partner:41559
TEST=Monitor cpufreq_* events in kernel debugfs and verify that the
     interactive timer task is rescheduled when CPU goes idle.

Change-Id: Iba78505712282a1825977756287ee10491c2852b
Signed-off-by: Ricky Liang <jcliang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/278094
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>