CHROMIUM: drm/nouveau/clk: Implement EMC scaling for GM20B
Add a minimum EMC frequency to each of the GM20B's P-states.
The EMC frequency at each state is determined using the same
method as in the legacy NVGPU driver, which is:

  EMC rate = 0.75 * (GPU bandwidth / CPU bandwidth) * GPU rate

For Tegra210/GM20B, this means the EMC rate should be ~3x the GPU
rate at each P-state.

Signed-off-by: Andrew Bresticker <abrestic@chromium.org>

BUG=chrome-os-partner:41674
TEST=EMC scaling on Smaug.

Change-Id: I36d70fee41031564a0ff8e618503984b21830317
Reviewed-on: https://chromium-review.googlesource.com/282506
Reviewed-by: Puneet Kumar <puneetster@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>