ANDROID: cpu_power: Avoids race condition when the task exits.
When the task is terminated, the cpu_power for that particular task is
added to the terminated tasks. cpu_power is set to ULLONG_MAX at this
point to avoid double accounting of the power. It is possible that before the
task releases all the resources, cpu reschedules the task or a timer
interrupt is fired. At this point we will try to add the additional time
to the process, which will cause the accounting to be skewed. This
avoids the case where we change the cpu_power when it is already set to
ULLONG_MAX.

BUG=b:24233083
TEST=manual

Original-Change-Id: I405733725d535b0a864088516bf52fa3638ee6aa
Change-Id: I461f3f78bbcd84e69a4c340bd9a45701f33f453d
Signed-off-by: Jin Qian <jinqian@google.com>
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/301155
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>