cros_ec: Re-add "mkbp" as possible vboot context type
We may have been too zealous earlier in our efforts to eliminate
"mkbp" from the ChromeOS EC lexicon. This re-adds "mkbp" so it
will be checked in addition to "cros-ec" just in case we really
need to support old firmware versions with the new kernel.

It would be nice to eventually eliminate use of the term "mkbp."
We can revisit this after kernel 3.18.

BUG=chrome-os-partner:21097
BRANCH=none
TEST=compiled 3.18 kernel for oak

Change-Id: I628288c28fe8605064b251821e819f2970fc9141
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/282763
Reviewed-by: Stefan Reinauer <reinauer@google.com>