UPSTREAM: tpm: Fix initialization of the cdev
When a cdev is contained in a dynamic structure the cdev parent kobj
should be set to the kobj that controls the lifetime of the enclosing
structure. In TPM's case this is the embedded struct device.

Also, cdev_init 0's the whole structure, so all sets must be after,
not before. This fixes module ref counting and cdev.

Cc: <stable@vger.kernel.org>
Fixes: 313d21eeab92 ("tpm: device class for tpm")
Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>
Reviewed-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>
Reviewed-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>
Tested-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>
Signed-off-by: Peter Huewe <peterhuewe@gmx.de>
(cherry picked from commit ba0ef85479c46a2ab354c2220bdb6152f7f4baf3)

BUG=chrome-os-partner:47234
TEST=verify no regression in TPM 1.2 systems

Change-Id: I97f7d4cf40f55e306d37423366686df53eaa3839
Signed-off-by: Bryan Freed <bfreed@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/316569
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>