CHROMIMU: power: reset: max77620: always hook pm_power_off when DT has power off property
In PSCIv1.0, it supports system power off function and hooks the
pm_power_off for it. But the PSCI firmware may not implement this which
will cause the system power off function fail.

So we fix it by changing the scenario as if the DT has power off
property, then we overide the pm_power_off for it.

BUG=chrome-os-partner:40741
TEST=run "reboot -p"

Change-Id: I82cf6bb87f255f5710acbdc25a19019987f91fa6
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/284153
Reviewed-by: Benson Leung <bleung@chromium.org>