CHROMIUM: drm/nouveau: Add GEM_SET_INFO ioctl
This patch introduces a new GEM_SET_INFO ioctl, which is similar to
GEM_SET_TILING, but generalized to let the userspace client specify
the same set of GEM parameters as with GEM_NEW ioctl.

The purpose is primarily to add ability to configure GPU cacheability
of the buffer (added in further patch), but other settings might be
added later.

The new ioctl is designed to take the same drm_nouveau_gem_info struct
as GEM_INFO, update buffer parameters and return the resulting GEM
information for userspace to be able to confirm new settings.

BUG=chrome-os-partner:42636
TEST=Boot to UI (without GPU L2 enablement patch)

Change-Id: I6da5c6f1bca5b6633c56643358b3eaacc06ee6be
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/285755
Reviewed-by: Alexandre Courbot <gnurou@gmail.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>