CHROMIUM: drm/nouveau: mmu: Maintain PTE coherency after maps/unmaps
Although the L2 is yet enabled for GPU pages, but the PTE and PDE tables
are cached. So it's necessary to invalidate the L2 after altering the
tables with CPU (map_sg*) to keep the GPU view on them coherent, while
after modifying them with PRAMIN (unmap/map_at) it is necessary to flush
the cache to make CPU view coherent.

To achieve this, the patch extends the nvkm_interface with flush and
invalidate methods and forks the gk104 and gm107 ltc implementations
into gk20a and gm20b to implement them for appropriate GPUs.

BUG=chrome-os-partner:42859
TEST=UI working fine; no faults in sheets

Change-Id: Ibeb3ff6c3e27d113b0a41dd222798fa35d5514b5
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Signed-off-by: Vince Hsu <vince.h@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/287020
Reviewed-by: Puneet Kumar <puneetster@chromium.org>
Commit-Queue: Puneet Kumar <puneetster@chromium.org>
Tested-by: Puneet Kumar <puneetster@chromium.org>