CHROMIUM: drm/nouveau: Fix sgt mapping with length < table length
Current implementation of nvkm_vm_map_sg_table() iterates over pages of
single SG segment and increments pte counter. Then mmu->map_sg() is
called with (pte - len) as first PTE, but if somehow len ends up being
greater than num, we will end up with len not being equal to the number
of iterated pages, causing the first PTE to be calculated incorrectly.

This patch makes len be clamped to num at the beginning of the main loop
body, so the check for num can be removed from the loop iterating over
particular pages, while the variable can be decremented outside of the
loop. Moreover, there is no need to decrement pte inside the loop as
well, so the incrementation is moved after mmu->map_sg() is called and
done for all mapped pages in one go. This way we can handle the special
case correctly, while making the code simpler.

BUG=none
TEST=UI is working fine

Change-Id: Ic8fac3ef2afba10d235ec51813d98f6dde568138
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/287646
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>