UPSTREAM: regulator: max8973: Set VSEL regmap ops if DVS GPIO is not set
Use regmap helpers for get_voltage_sel and set_voltage_sel ops
if the DVS GPIO is not set.

The DVS GPIO allows on the fly selection of the VSEL register
from two choices. However, if it is not set, the VSEL register
will stay fixed and we can use the regmap ops. This allows use
of the *hardware_vsel* regulator APIs.

BUG=chrome-os-partner:45774
TEST=build and boot SMAUG EVT2

Change-Id: I74cef28898c9207e40ed7f5a6bda04cdbdf5f6ea
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Reviewed-on: https://chromium-review.googlesource.com/297932
Commit-Ready: Hao (Alex) Lei <alei@nvidia.com>
Tested-by: Hao (Alex) Lei <alei@nvidia.com>
Reviewed-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Hao (Alex) Lei <alei@nvidia.com>
Reviewed-by: Laxman Dewangan <dewangan.laxman@gmail.com>