rtc: armada38x: Align RTC set time procedure with the official errata
According to Armada38x functional errata FE-3124064, Write to RTC TIME register
may fail. As a workaround, issue a dummy write of 0x0 twice to RTC Status
register before writing to the RTC TIME register.
This is the updated implementation of the Errata that eliminates the need
of the long 100ms delay during the RTC set time procedure

BUG=chrome-os-partner:43527
TEST=test a loop of: 'date -s "2015-01-01 11:12:13"; hwclock -w; hwclock;'

Change-Id: I93ce35e68d8e379eb79a0ca1af13356ddf7bd536
Signed-off-by: Nadav Haklai <nadavh@marvell.com>
Reviewed-on: https://chromium-review.googlesource.com/289191
Reviewed-by: Grant Grundler <grundler@chromium.org>