CHROMIUM: thermal: fix pd_governor in race condition
There has race condition on the value "cdev->updated",
when it occured, the cooling state may not be update properly,
the state can't be tropped to 0 when the temperature
is lower than the trip value.
This patch force the pd_governor to update cooling
state in this condition to fix this issue.

BUG=chrome-os-partner:39762
TEST=heat and cool the system, the cooling state can
be updated properly.

Change-Id: Ib51b82da93d6bd9886e9ba69277475cf193028a9
Signed-off-by: Wei Ni <wni@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/291202
Tested-by: Wei Ni <wni.nvidia@gmail.com>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Puneet Kumar <puneetster@chromium.org>
Commit-Queue: Puneet Kumar <puneetster@chromium.org>