CHROMIUM: clk: tegra: fix the IRQ might be re-enabled in the DFLL resume code
The clk suspend/resume call is happened in the syscore stage which
doesn't allow the IRQ be enabled. But invoking the
pm_runtime_get/put_sync pair in the function might cause the IRQ being
enabled by accident, removing that to fix this issue and we don't really
need that here too. This can avoid the IRQ be enabled too early before
the noirq stage in the resume flow.

BUG=chrome-os-partner:40741,chrome-os-partner:43024
TEST=LP0 suspend/resume test and make sure the IRQ wasn't enabled in
     the syscore stage

Change-Id: If320050e044d8f54bd187f176327ce63c30b4cc9
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/291185
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>