CHROMIUM: drm/nouveau/mmu: use lazy mapping only for the memory needed by doing pushbuf
With lazy mapping, the memories are really mapped just before doing
pushbuf. But some memories are not allocated for doing pushbuf, so
they may not be really mapped due to the lazy mapping and then PTE
fault may happen if they are accessed.

This patch adds a "lazy" argument to the "nouveau_bo_vma_add" function,
so that it can be controlled that what memories should be applied the
lazy mapping, and what should be not.

BUG=chrome-os-partner:43736
TEST=Smaug can boot to UI and GPU runtime PM can work

Change-Id: I8f8994edd3b0b222f4945590cac82f4d450d38e1
Signed-off-by: Kary Jin <karyj@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/291570
Reviewed-by: Alexandre Courbot <gnurou@gmail.com>
Reviewed-by: Anatol Pomazau <anatol@google.com>
Tested-by: Anatol Pomazau <anatol@google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>