UPSTREAM: ASoC: ssm4567: Add sense support
The ssm4567 has sensing circuitry that can be used to monitor the current
and voltage on the speaker amplifier output has well as the VBAT input.
This data can be output over the I2S interface so it can be processed by a
DSP or similar.

This patch adds the sense capture output stream to the CODEC DAI as well as
DAPM widgets that ensure that the sensing circuitry is powered up when the
capture stream is active.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@kernel.org>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git
br: for-next
(cherry picked from commit dbe71b9d86ee77cf58a92657c43b0e48954dc62b)

BUG=chrome-os-partner:41280
TEST=with the whole series applied, audio works on SKL.
TEST=there are no regressions observed on BSW.

Change-Id: I79472b50519fe903371a82713817c33c2582bc1d
Signed-off-by: Fang, Yang A <yang.a.fang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/311252
Commit-Ready: James Ausmus <james.ausmus@intel.com>
Tested-by: Prathyushi Nangia <prathyushi.nangia@intel.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>