drm/tegra: dc: Don't add planes to the mask when inactive
On dpms, the atomic helper adds the primary plane to the
atomic state since it detects that a modeset is required.
On commit, the helper calls atomic_check, which then calls
the dc atomic_plane_check. This will add the plane to the
plane mask and cause an update to that plane on flush.

This patch avoids adding the plane to the mask such that
we won't update it.

BUG=chrome-os-partner:43877
TEST=Tested on smaug, planes not updated after dpms

Change-Id: Ifa43cb2b5721aac9977db0d52b05bd0231f398cc
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/293272
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Sean Paul <seanpaul@google.com>
Commit-Queue: Sean Paul <seanpaul@google.com>