HACK: arm64: dts: tegra210: set SW throttle temp to 70C
We set the Tskin-throttle at 43C, but it's not 100% effective
yet, because the Tskin coefficient data didn't be tuned fully,
the Tskin temperature is not right.
And if the Tskin is 43C, the CPU/GPU temperature is about 60C~70C,
so I hack the throttle temp to 70C to emulate the Tskin behavior.

BUG=chrome-os-partner:44169
TEST=the throttle will be triggered at 70C.

Change-Id: I5d58093f66e27601a3967ab420098125c0d27b1f
Signed-off-by: Wei Ni <wni@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/293141
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Wei Ni <wni.nvidia@gmail.com>
Commit-Queue: Wei Ni <wni.nvidia@gmail.com>