drm/nouveau: Use offset in gem_set_info ioctl
If the offset in the gem_set_info ioctl is nonzero, redo the virtual address
allocation with that offset.

This is to support the CUDA programming model, which requires that for a
particular buffer, the GPU virtual address must match the CPU virtual
address.

BUG=chrome-os-partner:42636
TEST=Built and boot kernel.  Ran some NVIDIA internal tests that try to map
a buffer to a specific address.

Change-Id: I8cbc730472de8048d1957c94cfe23cdfdf244351
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/293637
Reviewed-by: Tomasz Figa <tfiga@chromium.org>