CHROMIUM: soc/tegra: pmc: Support BootROM reset I2C commands
On Tegra210 and later, a set of I2C commands may be programmed
into the PMC scratch registers which will be executed by the
BootROM on reset.  This can be used to restore voltage rails to
safe levels on warm reset.

BUG=chrome-os-partner:42259
TEST=Configure a command block to write to VSD0, trigger a warm
(PMC) reset or watchdog reset, read VSD0 back in coreboot and
observe that it contains the value programmed in the command block.

Change-Id: I8054ec24093318134539935f370cbd730026d3fc
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/294062
Reviewed-by: Tom Warren <twarren@nvidia.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>