CHROMIUM: drm/tegra: Force dc update for psr panels
While DC is active, queued DCS packets are not sent until the end of the
next scanout. In non-continuous mode (psr), this can be a long time, which
might cause timeout.

To counteract this, we'll force a dc update every time we send a packet
while the dc is active and the panel supports psr.

Based on work by Mark Zhang <markz@nvidia.com> and
Kary Jin <karyj@nvidia.com>.

Change-Id: Ic0fbc265f84158e02d2645b1ef0c90114fd56772
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295492
Reviewed-by: Sean Paul <seanpaul@google.com>
Commit-Queue: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>