UPSTREAM: ASoC: topology: Disable use from userspace
Since the topology API is still in sufficient flux for changes to be
identified disable the use of the userspace ABI by adding #error
statements to the code, ensuring that nobody relies on the headers as
currently defined.  It is expected that this change will be reverted for
v4.3.

Signed-off-by: Mark Brown <broonie@kernel.org>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git
br: for-next
(cherry picked from commit 8c90503bf246bebb48caa5590d41df755ba08550)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=chrome-os-partner:41280
TEST=with the whole series applied, audio works on SKL.
TEST=there are no regressions observed on BSW.

Change-Id: Id29380f132f98f1536ffd3bc6a8200a480201d97
Reviewed-on: https://chromium-review.googlesource.com/310332
Commit-Ready: Vedang Patel <vedang.patel@intel.com>
Tested-by: Prathyushi Nangia <prathyushi.nangia@intel.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>