HACK: drm/tegra: Wait for fb to latch before returning to userspace
When an fb update is still pending and not latched the buffer
should not be returned to user space.  Also print a warning message.

This is hacky due to using ERR_PTR() to bypass NULL checks, but should
be safe because the pointer is not propagated further to any code that
would deferefence it.

Change-Id: I842f2f6a24a2fe5d7c4dee567321b221ebf7d0a4
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/294850
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Commit-Queue: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>