CHROMIUM: tegra: Use host1x activity monitoring for runtime pm
Previously, we would keep the tegradrm client powered on as long as the
channel is open.  The channel, however, could be sitting idle.  This
changes makes us rely on host1x to power us off if the channel is seen to
be idle.

Before this change, doing a channel_open will result in the falcon being
booted...and the tegra device will never be powered off again (until
channel_close, at which point we are presumably done with this tegra
device).  After this change, even though we did a channel_open, the device
may still get powered down even though we are not done with this tegra
device.  So, on resume, we have to boot the falcon again.

BUG=chrome-os-partner:42698
TEST=Built and boot kernel.  Played a youtube video in a browser.  Paused
the video, and see that nvdec gets powered off.  Make sure video can resume.

Change-Id: I1ed250c9bd5ffdada9947f18331fccec6bbb9c0f
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/297277
Reviewed-by: Tomasz Figa <tfiga@chromium.org>