FROMLIST: mmc: sdhci-pxav3: enable modifying MMC_CARD bit during card initialization
On Marvell Armada 38x SoC's the MMC_CARD bit in SD_CE_ATA_1 register must be
set to 0x1 when a MMC card is supposed to work in DDR mode, or when commands
CMD11, CMD14 and CMD20 are used.

This commit enables the above for all MMC cards by modifying the host
registers during card initialization. It is done by using init_card callback.

Signed-off-by: Marcin Wojtas <mw@semihalf.com>

BUG=chrome-os-partner:43600
TEST=none

Change-Id: I29f445003bc60cf60494608cac3cc2b840493485
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/300472
Reviewed-by: Kan Yan <kyan@chromium.org>