CHROMIUM: clk: tegra210: fix pllx dyn step calculation
The logic for calculating the input rate used when figuring out
the proper dynamic steps for pllx was incorrect. It is supposed to
be calculated using parent_rate / m but it was just using the parent
rate directly, therefore using the wrong step values.

BUG=chrome-os-partner:39779
TEST=verify that the proper input rate is used and the correct
values for stepa and stepb are used.

Change-Id: I9699f9a1d7897b72e54310c926c8183beef8d70d
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/295845
Reviewed-by: Bill Huang <bilhuang@nvidia.com>
Reviewed-by: Benson Leung <bleung@chromium.org>