CHROMIUM: soc/tegra: pmc: Write reboot cause to scratch registers
The reset code in PMC_RST_STATUS is not useful since firmware
will assert AP_SYS_RESET_L at each boot, thus resetting PMC_RST_STATUS
to POR.  For this reason, PMC_SCRATCH202[2:0] has been allocated to us
for storing the reboot cause.  Set it appropriately based on whether
the reboot was clean (BOOTREASON_REBOOT) or if it was due to a kernel
panic (BOOTREASON_PANIC).

Note that the contents of PMC_SCRATCH202 are random on a cold boot,
so a separate flag in the off-chip RTC is used to indicate whether
or not the current boot is warm and thus if PMC_SCRATCH202 should
be trusted.

BUG=chrome-os-partner:44606
TEST=Check that PMC_SCRATCH202 is updated on both a normal reboot
and a panic.

Change-Id: I7e4f6995ec3993352b85fdbf242a97668d8ce908
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295783
Reviewed-by: Furquan Shaikh <furquan@chromium.org>