CHROMIUM: LSM: Assume read-only root if device is NULL
The factory install shim does the following at init time:
 1. mounts /newroot it as tmpfs (/ is initramfs)
 2. setup files and mounts
 3. switch_root /newroot /sbin/init

In this scenario, mnt->mnt_sb->s_bdev is NULL (this makes sense
as there is no underlying device for tmpfs), and pr_info
causes a panic as it tries to access mnt->mnt_sb->s_bdev->bd_dev.

Fix this by assuming that assuming that the block device is
read-only, as this prevents the sysctl file from being created
and is therefore more secure.

BUG=chrome-os-partner:42543
TEST=build a image with image_type=factory_install,
     boot from that, no kernel panic

Change-Id: I0b709d7c501cdb96a141063f90fa57ad4fb18033
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295814
Tested-by: Hsu Wei-Cheng <mojahsu@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>