UPSTREAM: drm/i915: use pch backlight override on hsw too
Currently the difference between backlight control on HSW vs. BDW/SKL is
that on HSW we modify the duty cycle on the CPU register, and have the
hardware pass the changes on to the PCH registers. We still drive the
PCH PWM on both. While HSW and BDW use the same LPT PCH, BDW does not
pass these messages on to the PCH. Therefore on BDW we need to enable
the PCH override bit, and program the PCH directly. (On SPT PCH, this
mode is the default.) We could as well do this on HSW too, and in fact
I've been told this is what a certain other operating system does. So
use PCH backlight override on HSW too.

This simplifies some follow-up code, but it does have the danger of
breaking backlight on HSW machines. It should work, but mysterious are
the ways of backlight.

While at it, name the related backlight hooks according to the PCH
rather than the CPU for clarity.

Cc: Clint Taylor <clinton.a.taylor@intel.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>
Reviewed-by: Clint Taylor <Clinton.A.Taylor@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>

BUG=chrome-os-partner:43071
TEST=Enable normal mode - /usr/share/vboot/bin/set_gbb_flags.sh 0x0

Change-Id: I9f7206d2823707a14c4915313b71fca35c75633a
Link: git://anongit.freedesktop.org/drm-intel ref:drm-intel-next
(cherry picked from commit 437b15b8017e0d946453c10794b0c5d4591cf180)
Signed-off-by: Shobhit Kumar <shobhit.kumar@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/308577
Tested-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Joe Konno <joe.konno@linux.intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: James Ausmus <james.ausmus@intel.com>