BACKPORT: drm/i915: initialize backlight max from VBT
Normally we determine the backlight PWM modulation frequency (which we
also use as backlight max value) from the backlight registers at module
load time, expecting the registers have been initialized by the BIOS. If
this is not the case, we fail.

The VBT contains the backlight modulation frequency in Hz. Add platform
specific functions to convert the frequency in Hz to backlight PWM
modulation frequency, and use them to initialize the backlight when the
registers are not initialized by the BIOS.

v2: Fix SPT and VLV. Thanks to Clint for the VLV code.

Cc: Clint Taylor <clinton.a.taylor@intel.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>
Reviewed-by: Clint Taylor <Clinton.A.Taylor@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>

Conflicts:
	drivers/gpu/drm/i915/intel_panel.c
	(minor context change, initialize backlight_hz_to_pwm for vlv as
	well to resolve compiler warning)

BUG=chrome-os-partner:43071
TEST=Enable normal mode - /usr/share/vboot/bin/set_gbb_flags.sh 0x0

Change-Id: I13539f8ce4a72d87d1ab3f675615974f8f2d8d2a
Link: git://anongit.freedesktop.org/drm-intel ref:drm-intel-next
(cherry picked from commit aa17cdb4f8367875d3a53c41e004fc02e983e317)
Signed-off-by: Shobhit Kumar <shobhit.kumar@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/308579
Tested-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Joe Konno <joe.konno@linux.intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: James Ausmus <james.ausmus@intel.com>