CHROMIUM: irqchip: gic: fix build with CONFIG_GIC_NON_BANKED
The loop was removed in commit 0bace00d8231 ("CHROMIUM: irqchip: Add
ability to notify non per-cpu GICs"), so this hasn't compiled since
then. Change the 'continue' to a 'return NOTIFY_OK', as looks
reasonable, and reindent the function so that it is easier to see the
nesting in this function.

BUG=chromium:529437
TEST=build ARM32 with CONFIG_GIC_NON_BANKED

Change-Id: Ia8b4ebc7f9dfef5453dbca2ddf07d7bea81a73cc
Fixes: 0bace00d8231 ("CHROMIUM: irqchip: Add ability to notify non per-cpu GICs")
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/298161
Reviewed-by: Ajay Nandakumar <anandakumarm@nvidia.com>
Reviewed-by: Anatol Pomazau <anatol@google.com>