CHROMIUM: input: gpio_keys: handle the missing key press event in resume phase
The GPIO key press event might be missed in the resume phase, if the key
had been released before the system had been resumed to the stage that it
could capture the press event. So we simulate the wakeup key press event
in case the key had been released by the time we got interrupt handler
to run.

BUG=chrome-os-partner:43795
TEST=wake up from LP0 by the very short time clicking the power key;
     the system should able to wake up and the UI comes up too.

Change-Id: I79dbb1c48a41c6a435f7e147a6c769c0ef0e7daf
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/296285
Commit-Ready: Joseph Lo <yushun.lo@gmail.com>
Tested-by: Joseph Lo <yushun.lo@gmail.com>
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>