UPSTREAM: soc: mediatek: Move the initial setting of pmic wrap interrupt before requesting irq.
The watchdog may not be initialized by the bootloader, even if the rest
of the pwrap is. Move the watchdog initialization out of pwrap_init() to
make sure the watchdog is always initialized and not only when the pwrap
is uninitialized.

Signed-off-by: Henry Chen <henryc.chen@mediatek.com>
Acked-by: Sascha Hauer <s.hauer@pengutronix.de>
Signed-off-by: Matthias Brugger <matthias.bgg@gmail.com>
(cherry picked from commit bbe379fc7ca5cae09b1bc44fe5ee4173ad883fbf
 https://github.com/mbgg/linux-mediatek)
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=build & boot oak-rev2 to UI

Change-Id: I01c1e74ad767ec0c70c0508ee9058fdcdc479367
Reviewed-on: https://chromium-review.googlesource.com/302949
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>