CHROMIUM: config: Configure builds to include hid-quickstep driver
The hid-quickstep driver is required for a touchpad/screen latency
device to work on ChromeOS devices.  A new version is in the works now
but in the meantime Google engineers and some partners need to be able
to easily run it in the old style.  This CL simply enables the
driver for 3.18 builds like it is built in the other versions of the
kernel.

BUG=chromium:530719
TEST=built and the .ko file is found on the device, so it built the
module and included it in the image.

Change-Id: Iae99c51d58549909460b42469ac4e2e79fcc3d39
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/299410
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>