CHROMIUM: usb: gadget: tegra_xudc: Disable U1
There is a U1->U2 inactivity timer in HW which is active in U1 state,
and the state machine automatically transitions to U2 when that timer
expires. Ideally this timer should be reset and stopped when U1 exit is
in progress; this is not happening, which causes the state machine to
abruptly change to U2 when U1 exit LFPS handshake is in progress.

This known issue will happen when transferring large files (e.g. 3.8GB)
from some XHCI port to device where incorrect state machine transition
is observed: U0->Inactive->RxDetect->U0.

So we're not able to enable both U1 and U2. Disable U1 only since major
power saving comes from U2.

BUG=none
TEST=Smaug XUSB device mode works normally

Change-Id: I35d0592096822ce81c8de5e1988686344183f3d2
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/299563
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>