FROMLIST: mmc: block: Add new ioctl to send multi commands
Certain eMMC devices allow vendor specific device information to be read
via a sequence of vendor commands. These vendor commands must be issued
in sequence and an atomic fashion. One way to support this would be to
add an ioctl function for sending a sequence of commands to the device
atomically as proposed here. These multi commands are simple array of
the existing mmc_ioc_cmd structure.

The structure passed via the ioctl uses a __u64 type to specify the number
of commands (so that the structure is aligned on a 64-bit boundary) and a
zero length array as a header for list of commands to be issued. The
maximum number of commands that can be sent is determined by
MMC_IOC_MAX_CMDS (which defaults to 255 and should be more than
sufficient).

Signed-off-by: Seshagiri Holi <sholi@nvidia.com>
Cc: Arnd Bergmann <arnd@arndb.de>
Cc: Grant Grundler <grundler@google.com>
Cc: Olof Johansson <olofj@chromium.org>
[ jonathanh@nvidia.com: Rebased on linux-next from v3.18. Changed
  userspace pointer type for multi command to be a u64. Renamed
  from combo commands to multi commands. Updated patch based upon
  feedback review comments received. Updated commit message ]
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>

[
$ pwclient list -w jonathanh@nvidia.com
Patches submitted by Jon Hunter <jonathanh@nvidia.com>:
ID      State        Name
--      -----        ----
7112081 New          [RFC] mmc: block: Add new ioctl to send combo commands
7117131 New          [RFC] mmc: block: Add new ioctl to send combo commands
7147241 New          mmc: block: Add new ioctl to send multi commands
7160441 New          [V2] mmc: block: Add new ioctl to send multi commands
7176481 New          [V3] mmc: block: Add new ioctl to send multi commands
$ pwclient git-am 7176481
... backport/merge by hand.
]

BUG=chrome-os-partner:43250
TEST=compile and boot kernel. RPMB not tested by me.

Change-Id: Ie995490ecb0f3937c9588661d1a637988a648059
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/299800