CHROMIUM: usb: xhci-tegra: enable wakeup for both hubs
Set wakeup capable for both USB2.0 and 3.0 hubs in XUSB, so that XUSB
will not be forced out of ELPG every time when either hub is called with
usb_suspend() from system suspend where choose_wakeup() checks if the
USB device is autosuspended with wrong (not matching) wakeup setting.

This will enhance suspend completion time if XUSB is already in ELPG.

This also fixes a bug in tegra_xhci_suspend() where return value is not
initialized when XUSB is already runtime suspended.

BUG=none
TEST=XUSB works across a few LP0 loops

Change-Id: I189c09bdcca1333079b8176ef094227a296b9558
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/299624
Tested-by: Joseph Lo <josephl@nvidia.com>
Reviewed-by: Jim Lin <jilin@nvidia.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>