CHROMIUM: drm/tegra: Don't take a vblank reference in atomic_begin
If the crtc is off, and the state to be committed includes dpms + an
event, we'll end up failing the vblank_get in atomic_begin. This will
cause a ref mismatch on the next vblank when we attempt to put the
ref.

We already do proper referencing in tegra_atomic_wait_for_flip_complete,
so doing it in atomic_begin is redundant anyways.

Change-Id: I752edac9e3340e5ee2bff6d51732b059cf12c748
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/299908
Commit-Ready: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>