CHROMIUM: iommu: Add support for out of band flushing
This patch adds a new "flush" callback to iommu_ops, which is supposed
to perform any necessary flushes within given IOMMU domain to make any
changes to mappings of given area [iova; iova + size) be reflected to
IOMMU clients.

The purpose is to let IOMMU drivers skip page-by-page flushes and
replace it with one flush of full address range on devices which support
it.

BUG=chrome-os-partner:45017
TEST=Boot to UI; launch some apps

Change-Id: I243dea907e1ad145153282efe34e8af9f49730ca
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/299993
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Tested-by: Vince Hsu <vince.h@nvidia.com>
Reviewed-by: Vince Hsu <vince.h@nvidia.com>
Reviewed-by: Puneet Kumar <puneetster@chromium.org>