CHROMIUM: drm/nouveau: Fix potential races with runtime PM
gem_unmap_work() and nouveau_gem_pushbuf_queue_process_work() run
asynchronously with respect to the Nouveau ioctls and thus are not
guaranteed to run with the GPU powered up.  Grab a runtime PM reference
in both cases to ensure that we don't enter runtime suspend while
these functions are running.

BUG=chrome-os-partner:44914,chrome-os-partner:44012
TEST=Reboot and suspend/resume cycles on Smaug.

Change-Id: I8bdc7f6b880ac0febf65c7929a2a2021a840b49c
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/300560
Tested-by: Joseph Lo <josephl@nvidia.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>