UPSTREAM: ASoC: rt5645: Add jd_invert for Broadwell
Broadwell can not triger the IRQ falling and rising simultaneously,
so it can not detect jack-in and jack-out simultaneously.
We add a flag "jd_invert" to platform data. If this flag is set,
codec IRQ will be set to invert that forces IRQ as pulse when jack-in
and jack-out.

BUG=chrome-os-partner:40164
TEST=test audio on broadwell platform

Signed-off-by: John Lin <john.lin@realtek.com>
Signed-off-by: Bard Liao <bardliao@realtek.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 917536aeb88d34e06c1353b0dd144f0987bb66bd)
Signed-off-by: Wonjoon Lee <woojoo.lee@samsung.com>

Change-Id: I8fb290629d57af9f44dee97374b353ba49accdd4
Reviewed-on: https://chromium-review.googlesource.com/304461
Commit-Ready: Wonjoon Lee <woojoo.lee@samsung.com>
Tested-by: Wonjoon Lee <woojoo.lee@samsung.com>
Reviewed-by: Ben Zhang <benzh@chromium.org>