ANDROID: sched: cpufreq: Update cpu_power field in the task_struct
cpu_power has been added to keep track of amount of power each task is
consuming. cpu_power is updated whenever stime and utime are updated for
a task. power is computed by taking into account the frequency at which
the current core was running, and the current (uA) consumed by the cpu.

BUG=b:24233083
TEST=manual

Original-Change-Id: Ic535941e7b339aab5cae9081a34049daeb44b248
Change-Id: I233bc648e4f537b6807765a7695f2897f47c783b
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>
Signed-off-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/301153
Reviewed-by: Stephen Barber <smbarber@chromium.org>