HACK: usb: Serialize probe of Tegra xHCI host and XUDC
There is an issue in the power-ungating sequence of the XUSB device
controller (XUDC) such that an interleaving of this sequence with an
MMIO access to the xHCI registers causes a hard hang (in particular,
it appears that the XUSB_SS reset is spuriously asserted).  To work
around this until the root cause is identified, serialize the power-up
and initialization of the xHCI host and XUDC with respect to one another.

NOTE: It's theoretically possible to hit this issue during runtime PM or
suspend/resume, but it has yet to be observed.  This patch only attempts
to address the issue at boot time.

BUG=chrome-os-partner:43951
TEST=Reboot cycles on Smaug.

Change-Id: I80475668063c1d840e2cb7347d601c5c4a98fba2
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Signed-off-by: Jim Lin <jilin@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/301991
Reviewed-by: Puneet Kumar <puneetster@chromium.org>