ANDROID: usb: phy: update dynamically dual role sysfs nodes
When the USB connection state is updated, read again the ->writeable()
property of the driver to be able to adjust dynamically the capabilities
available to the accessory plugged on the USB port.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=b:24137353
BUG=b:24645641
TEST=manual, update the Smaug driver to send various values from its
->writeable() callback depending of the type of USB accessory plugged.
Check the content of /sys/class/dual_role_usb/otg_default/ and the UI
content.

Change-Id: I3aee540383bfe8c3837ee890967b5a8bc9be96c9
Reviewed-on: https://chromium-review.googlesource.com/302492
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>