CHROMIUM: drm/nouveau: Flush unmap queue on remap
In the fixed address mapping case, when we try to remap a buffer object
to a fixed address, we may run into a problem where a previous buffer object
unmap is still sittingin the unmap queue, and we immediately want to map
a different (or perhaps the same) buffer object to the same address.  This
will fail because the unmap hasn't happened yet, so if we fail the remap,
flush the unmap queue and try again.

We could simply just flush the unmap queue on every fixed address remap, but
I thought that would negate the benefits of deferring the unmaps.

Also refactored this remapping code a bit to fix some indentations that were
getting out of hand.  The only functional difference should be that if we
fail the call to nouveau_bo_vma_add_offset(), flush the unmap queue and try
again.

BUG=chrome-os-partner:42636
TEST=Built and boot kernel.  Perform fixed address mappings via compute
tests, with instrumentation at the retry point, and see that it kicks in
and handles the remap properly.

Change-Id: I1efa5f5a871621b54cbc4ebec671e66b53807376
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/302584
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>