CHROMIUM: drm/nouveau: Cancel deferred map only if not yet mapped
In nouveau_gem_object_unmap(), check to see if the vma has actually been
mapped, and only remove it from the deferred map list if it hasn't yet been
mapped.  If it has been mapped already, it won't be on the deferred map
list.  This saves us a lot of walks down the deferred map list.

BUG=chrome-os-partner:45411
TEST=Run GLES3.1 test and see test pass.

Change-Id: I04a148041564d2e313aa5d8dfe7a8d5a840fcb76
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/301680
Reviewed-by: Tomasz Figa <tfiga@chromium.org>