FROMLIST: PM / Domains: add setter for dev.pm_domain
Adds a function that sets the pointer to dev_pm_domain in struct device
and that warns if the device has already finished probing. The reason
why we want to enforce that is because in the general case that can
cause problems and also that we can simplify code quite a bit if we can
always assume that.

This patch also changes all current code that directly sets the
dev.pm_domain pointer.

BUG=chrome-os-partner:46491
TEST=boot glados

Signed-off-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
(am from https://patchwork.kernel.org/patch/7328141/)
Signed-off-by: Derek Basehore <dbasehore@chromium.org>

Change-Id: I2eaf04084337c9a9b6f7f188f94d7ebada8dbb57
Reviewed-on: https://chromium-review.googlesource.com/305814
Commit-Ready: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>