FROMLIST: PM / sleep: Go direct_complete if driver has no callbacks
If a suitable prepare callback cannot be found for a given device and
its driver has no PM callbacks at all, assume that it can go direct to
complete when the system goes to sleep.

The reason for this is that there's lots of devices in a system that do
no PM at all and there's no reason for them to prevent their ancestors
to do direct_complete if they can support it.

BUG=chrome-os-partner:46491
TEST=suspend_stress_test on glados

Signed-off-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
(am from https://patchwork.kernel.org/patch/7328171/)
Signed-off-by: Derek Basehore <dbasehore@chromium.org>

Change-Id: Ic3d6cd732f6c64ef34f28e065454f690eed4e6f8
Reviewed-on: https://chromium-review.googlesource.com/305815
Commit-Ready: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>