UPSTREAM: ASoC: Intel: Skylake: Add pipe and modules handlers
SKL driver needs to instantiate pipelines and modules in the DSP.
The topology in the DSP is modelled as DAPM graph with a PGA
representing a module instance and mixer representing a pipeline
for a group of modules along with the mixer itself.

Here we start adding building block for handling these. We add
resource checks (memory/compute) for pipelines, find the modules
in a pipeline, init modules in a pipe and lastly bind/unbind
modules in a pipe These will be used by pipe event handlers in
subsequent patches

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git
br: for-next
(cherry picked from commit e4e2d2f45284f620eeb9eea25570838b2a42c968)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=chrome-os-partner:41280
TEST=with the whole series applied, audio works on SKL.
TEST=there are no regressions observed on BSW.

Change-Id: I4498b8f1c530654ecd35653f99422e919d2e82dd
Reviewed-on: https://chromium-review.googlesource.com/310341
Commit-Ready: Vedang Patel <vedang.patel@intel.com>
Tested-by: Prathyushi Nangia <prathyushi.nangia@intel.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>