drm/nouveau: make client's notify array dynamic
The number of notify objects per channel varies by several factors, and
now with the error notifier it depends also linearly on the number of
channels. Instead of the fixed 32, reallocate the notify object array to
twice the previous space when adding a new element if it's full. The
first allocation size is 16.

Note that this doesn't grow boundless even though the array never
shrinks and is only completely deleted at dtor(). Dead notify objects
are replaced with NULL and those slots are reused when new ones are
allocated, as before.

Change-Id: I970e7e0f334e3ba8626ab07b8e6d3569339f39bc
Signed-off-by: Konsta Holtta <kholtta@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/304520
Reviewed-by: Vince Hsu <vince.h@nvidia.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>