CHROMIUM: arm64: dts: mt8173-oak-rev4: Add dts and make default
oak-rev4 is very similar to oak-rev3, with just a few minor
differences.

The only difference that has already been committed is that
PANEL_LCD_POWER_EN is changed on oak-rev4 from GPIO 29 to 88.
Set oak-rev4 as the new default, and move the old value to
oak-rev3.dtsi.

Signed-off-by: Yidi Lin <yidi.lin@mediatek.com>

BUG=chrome-os-partner:46540
TEST=build and boot oak-rev3 and -rev4 to UI.

Change-Id: I586939dc97bd1f52cd195318ece721d99090ea2f
Reviewed-on: https://chromium-review.googlesource.com/301261
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>