Revert "CHROMIUM: jdi-lpm102a188a: fix unbalanced regulator disable on shutdown"
Recent touchscreen firmware changes have made it possible to save nearly
as much power in S3 as when we forcibly reset the device and turned off
its power supplies.  As those techniques occasionally introduced noise
on the sensor its best to leave it powered and running if power is
acceptable.

Signed-off-by: Todd Broch <tbroch@chromium.org>

BUG=chrome-os-partner:43346
TEST=measure power in S3 w/ new touchscreen fw and its nearly as low as
when its held in reset and power rails are turned off.

This reverts commit 57f234003351eb3239868f2111e1bac4c031b325.

Change-Id: I05b17135a5f387f936c779e485136bc1def87cba
Reviewed-on: https://chromium-review.googlesource.com/306611
Commit-Ready: Todd Broch <tbroch@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>