Revert "CHROMIUM: jdi-lpm102a188a: Rework regulator usage for S3 support"
This reverts commit a0cd7b02961eda886667d9167a0d11007893a13c.

Recent touchscreen firmware changes have made it possible to save nearly
as much power in S3 as when we forcibly reset the device and turned off
its power supplies.  As those techniques occasionally introduced noise
on the sensor its best to leave it powered and running if power is
acceptable.

Signed-off-by: Todd Broch <tbroch@chromium.org>

BUG=chrome-os-partner:43346
TEST=measure power in S3 w/ new touchscreen fw and its nearly as low as
when its held in reset and power rails are turned off.

Change-Id: Id6f0336ff3de7676b7cc093dbe5bb95b6a4fbd81
Reviewed-on: https://chromium-review.googlesource.com/306612
Commit-Ready: Todd Broch <tbroch@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>