CHROMIUM: drm/nouveau/clk: change EMC to GM20B clock rate mapping
Currently the EMC rate is rounded down from 3x the GPU rate at each
P-state, for some GPU use cases it is found that EMC rate is not high
enough to be underflow free, this might not be a problem when we are
doubling the DC bandwidth request, but recently we have removed that
hack in DC driver (since we have LA/PTSA programming for DC), so making
sure GPU is requesting high enough EMC frequency is important. This
change make the minimum EMC frequency at each state to be rounding up
from 3x the GPU rate at each P-state.

BUG=chrome-os-partner:46932
TEST=adb shell cat /d/dri/1/dc.0/regs | grep UFLOW while playing a game
     such as Epic Citadel to see if you can see DC underflow, it should
     not.

Change-Id: Ia937a62d0372cb687645cb5050e4d878d651787e
Signed-off-by: Bill Huang <bilhuang@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/309120
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>