CHROMIUM: zram: lz4: fallback to vmalloc allocation
zram lz4 compression algorithm requests order-3 page allocations.
When system memory is fragmented, some of the order-3 allocations
start failing, therefore fallback to vmalloc which will usually
result in a couple of order-0 allocations and hence will work even
if memory is fragmented.

lz4 compression algorithm is using this allocated buffer as a
hashtable and so should not fail even if underlying buffer is
allocated from vmalloc instead of kmalloc.

BUG=chrome-os-partner:47247
TEST=run Epic Citadel, Fruit Ninja, multi-tab chrome browsing and
     observe for allocation failures in dmesg

Change-Id: Ie3821299424fb0a3e2ef593430d2bbc8baef3b4a
Signed-off-by: Maneet Singh <mmaneetsingh@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/310535
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>