CHROMIUM: xhci: Clear device initated resume variables once we are in U0
Resuming a usb2 port from U3 to U0 is done slightly differently
depending on if the resume is host initiated or device initiated.

In common we have U3 -> RESUME -> U0
We need to stay in RESUME for 20ms before we are allowed to move to U0

for host initiated resume:
- write RESUME state
- msleep(20)
- write U0 state

for device initiated resume:
- get a port status event saying we are in RESUME
- set a "port in resume" flag and tag current time
- kick root hub timer that will start hub polling port status

- when hub reads port status, device is in RESUME, port in resume flag
  is set and current time > tagged time+20ms THEN:
- write U0 state to port

This is prone to races:
if we in the middle of a host initiated resume, while msleep(20) where
spinlock is disable handle a port event for the same port. It will catch the
port in RESUME state, believing it was a device initiated resume,
set "port in resume" tag and tag current time.  We will however not catch the
port in RESUME state again when hub reads port status as host initiated resume
has already moved the port state to U0. In this case the "port in resume" flag
is always set, and time tag is set. This confuses the port status and ends up
trying to resume a port in U0 in a forever loop. Fix this for now by always
clearing the device initiated resume parameters once port is in U0

This patch will be upstreamed after release 4.4-rc1

BUG=chrome-os-partner:47633
TEST=Verify auto suspend after manual runtime resume

Change-Id: I6be414240f01b5ba624de60a2cac4ce3f4028035
Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
Signed-off-by: Jenny TC <jenny.tc@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/312583
Commit-Ready: Jenny Tc <jenny.tc@intel.com>
Tested-by: Jenny Tc <jenny.tc@intel.com>
Reviewed-by: Julius Werner <jwerner@chromium.org>