CHROMIUM: mmc: sdhci-tegra: Warn on register accesses while clock gated
If a Tegra210 SDHCI register is accessed while the core is clock gated,
it will hang the bus, causing a watchdog reset.  We've seen a few
reports of watchdog resets immediately following SD bus errors, leading
to speculation that there might be an improper register access somewhere.
Add a few new checks that will log a warning + backtrace if this happens.

BUG=chrome-os-partner:45478
TEST=manual

Change-Id: If9e3d719ad0935d25d575aa7054ebbcade2d4236
Signed-off-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/311856
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>