CHROMIUM: drm/tegra: Add mutex around contexts list
In rare cases, there is a crash in tegra_drm_file_owns_context() while
iterating the file->contexts list.  This is because while this iteration
is happening, someone opens or closes a channel, which can modify the list.

Protect access to this contexts list with a mutex.

BUG=chrome-os-partner:47827
TEST=Built and boot on smaug. Repeatedly run
cts2.testStillPreviewCombination (which previously can repro this crash
after 50 or so iterations).

Change-Id: Iee7dad4aed49ad1a93dea798c9cf60cfeb4305f9
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/313001
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>