CHROMIUM: usb: gadget: tegra_xudc: Don't poll for port reset completion
Quickcharge chargers are generating Port Reset Change (PRC) events upon
connection which never complete.  Since the XUDC appears to generate
PRC events on both 1-to-0 and 0-to-1 transitions of the Port Reset (PR)
bit, we can avoid polling for completion of the port reset by ignoring
the 0-to-1 transition.  The port reset will then be handled on the PRC
event generated by the 1-to-0 transition.  Do the same for WRC/WPR to
avoid a similar situation there.

BUG=b:25699970
TEST=Connecting a QC2.0 charger to Smaug no longer hangs the system.
ADB, MTP, etc. still work.

Change-Id: Ie059f827b8ff4675ff69148bc63358c8a5476c78
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/312866
Reviewed-by: Jim Lin <jilin@nvidia.com>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>