CHROMIUM: usb: gadget: tegra_xudc: Fix request dequeue logic
The current implementation of __tegra_xudc_ep_dequeue() does not
correctly detect if the TD corresponding to the request to be dequeued
is in progress.  In particular, the check for if the TR dequeue pointer
is before the TD is wrong and the logic in trb_processed() does not
correctly handle several corner cases.  This results in the hardware
continuing to process TDs even after the corresponding request has been
dequeued.

To fix this, introduce two new helpers trb_{in,before}_request() (which
determine if a TRB is part of a request or before a request in the TR,
respectively) and clean up the logic in __tegra_xudc_ep_dequeue().

As before, there are three possible cases for dequeueing an active TD
within the TR, depending on the location of the TR dequeue pointer:
  1) Within the TD and the hardware owns the TRB (i.e. the endpoint's
     Cycle State matches the Cycle bit of the TRB): the TD is aborted
     and the TR is compressed.
  2) Before the start of the TD and the hardware owns the TRB: the TD is
     removed from the TR and the TR is compressed.
  3) Beyond the end of the TD or the hardware does not own the
     TRB (i.e. there are no active TDs): TD has already completed.

BUG=b:24941427
TEST=Send files over MTP to Smaug from a Mac OS X host and observe that
canceling the transfer no longer causes a disconnect.

Change-Id: I094f7d2f847c3487da5cb51cf069482e60001226
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/315166
Reviewed-by: Benson Leung <bleung@chromium.org>