CHROMIUM: usb: gadget: tegra_xudc: Fix port resume
In tegra_xudc_resume_device_state(), the PLS value is incorrectly
determined so the if statement will never be true.  Thus, the port
will never be directed to U0.   Fix the statement to correctly get PLS
and also set PLS to U0 instead of RxDetect to actually resume the port.

BUG=none
TEST=Build and boot on Smaug.

Change-Id: Idee87972e2101eb7545b9e8ca50cfad24a44ee10
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/313855