CHROMIUM: usb: gadget: tegra_xudc: correctly re-enable EP
When enabling an EP, we missed to check if the EP is already enabled.
When re-enabling the same EP, the "nr_enabled_eps" count will be
incorrectly increased by 1. This has caused unbalanced nr_enabled_eps
which is used as an indicator to switch from Configured state to
Addressed state when disabling endpoints in __tegra_xudc_ep_disable().
The result will be that XUDC stays in Configured state even if it should
be in the Address state.

Fix this by checking EP status first and if it's already enabled,
first disable the EP before enabling it.

BUG=none
TEST=Build and boot on Smaug

Change-Id: I797440e808e71a36e5669fbf554371fa7fe08297
Signed-off-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/314545
Commit-Ready: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Jim Lin <jilin@nvidia.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>