CHROMIUM: usb: gadget: rndis: fix print format warning
'sizeof' has 'size_t' type, so on 64-bit platform '%u' is the incorrect
size as a printf format. Use '%zu' which is valid on both 32-bit and
64-bit platforms.

Note: the pr_err with the faulty format is from patch
'c5ecbca usb: Add support for rndis uplink aggregation' which is not upstream.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:48215
TEST=build with CONFIG_USB_CONFIGFS_RNDIS.

Change-Id: I4ad7714894033afe7c72b31e40bc4804552c57ab
Reviewed-on: https://chromium-review.googlesource.com/315135
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>