UPSTREAM: ASoC: Intel: Skylake: Add support for active suspend
Some of the usecases can be marked as 'ignore_suspend' by
machine. For these on suspend we should keep audio controller
ON by saving the state and not suspending the device

For this we need to maintain a counter for these streams and be
active on suspend when such a stream is opened.

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git
br: for-next
(cherry picked from commit 4557c305d4fc9356563a1d41fa6fe29e494f0460)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=chrome-os-partner:48270
TEST=with the whole series applied, audio works on SKL.
TEST=there are no regressions observed on BSW.

Change-Id: I3db774c1f0f07050425f84112ff4915df1b4e205
Reviewed-on: https://chromium-review.googlesource.com/317443
Commit-Ready: Radhakrishna Sripada <radhakrishna.sripada@intel.com>
Tested-by: Sathyanarayana Nujella <sathyanarayana.nujella@intel.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>