CHROMIUM: ASoC: rt5677: Handle registers used by hotword DSP fw
Some registers are modified by the hotword DSP firmware and
the codec driver. When CONFIG_SND_SOC_RT5677_HOTWORD is
enabled, they need to be marked as volatile to disable
regcache so that changes made by the DSP firmware are visible
to the codec driver. When the hotword config is enabled but
the user decides not to use hotwording during suspend, these
registers need to be manually saved/restored at suspend/resume
because the codec loses power and resets when there is no
active stream at suspend. regcache doesn't save/restore them
in this case because they are volatile.

BUG=chrome-os-partner:48039
TEST=normal playback/capture, hotwording during suspend works on Ryu

Change-Id: Ife195dc689ed0152b1397e1f87e1c1dde481ee74
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/316724
Reviewed-by: Anatol Pomazau <anatol@google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>