CHROMIUM: PM / OPP: Fix OOB access in _opp_add_dynamic()
When iterating through device's OPP list we may not find frequency that
is higher than the one that we are trying to insert, in which case opp
pointer is invalid and we should not be trying to reference it.

Fix it by using a temporary for iterating and saving the valid pointer
to be used later.

BUG=chromium:568656
TEST=Build and booted on Smaug

Change-Id: I6eb709142c6d252b465613baf3b22b42b99b1cbd
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/318251
Tested-by: Alexander Potapenko <glider@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>