CHROMIUM: ASoC: rt5677: Don't overflow hotword stream dma_area
If the space left in the dma_area buffer is less then one
period, stop copying from DSP and wait for userspace to read,
so that we don't overflow. The DSP has a 4sec buffer which is
usually much larger than the dma_area buffer.

This fixes the issue where userspace sees repeated partial
"OK Google" in the hotword stream caused by buffer overflow.
And it inserts "OK Google" into the search query. This
usually happens when waking from suspend.

BUG=chrome-os-partner:48039
TEST=20 hotword wake from suspend on Ryu without error

Change-Id: I0f4fae1e643d87bd6eb0122e19e192aa1020fc21
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/318892
Reviewed-by: Anatol Pomazau <anatol@google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>