UPSTREAM: driver core / PM: Add PM domain callbacks for device setup/cleanup
If PM domains are in use, it may be necessary to prepare the code
handling a PM domain for driver probing.  For example, in some
cases device drivers rely on the ability to power on the devices
with the help of the IO runtime PM framework and the PM domain
code needs to be ready for that.  Also, if that code has not been
fully initialized yet, the driver probing should be deferred.

Moreover, after the probing is complete, it may be necessary to
put the PM domain in question into the state reflecting the current
needs of the devices in it, for example, so that power is not drawn
in vain.  The same should be done after removing a driver from
a device, as the PM domain state may need to be changed to reflect
the new situation.

For these reasons, introduce new PM domain callbacks, ->activate,
->sync and ->dismiss called, respectively, before probing for a
device driver, after the probing has completed successfully and
if the probing has failed or the driver has been removed.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Ulf Hansson <ulf.hansson@linaro.org>
Reviewed-by: Kevin Hilman <khilman@linaro.org>
Acked-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit e90d5532773e2bcccc538dd346b9fc3482cd700c)
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
[djkurtz: Resolved context conflict in drivers/base/dd.c:really_probe()
 because chromeos-3.18 already has commit 52cdbdd49853 ("driver core:
 correct device's shutdown order").
]

BUG=chromium:571567
TEST=build & boot to UI on oak-rev4
TEST=build & CQ test smaug, strago

Change-Id: I6e651802c05a51ef772a4b6228a423d4ef03a475
Reviewed-on: https://chromium-review.googlesource.com/319759
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>