FROMLIST: ASoC: hdac_hdmi: create dais based on number of streams
A stream is mapped to a converter. So based on the converters
queried, dais are created.

The streams can be dynamically routed to any converter. For
now it is mapped statically. The dynamic mapping of stream
to converter will be added when required.

Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
(am from https://patchwork.kernel.org/patch/7945831/ )
Signed-off-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>

BUG=chrome-os-partner:48269
TEST=with the whole series applied, audio works on SKL.
TEST=there are no regressions observed on BSW.

Change-Id: I2c6ff4276d206e85c5cac8a817060393bb8629a6
Reviewed-on: https://chromium-review.googlesource.com/316205
Commit-Ready: Radhakrishna Sripada <radhakrishna.sripada@intel.com>
Tested-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>