CHROMIUM: iio: cros_ec: Use proper response size on the data path.
For simplicity sake, all responses from EC were expected to be up
to max_response size (120 bytes with i2c).
The drawback is the host would retrieve 130 bytes from the EC for
every response to sensor commands.
This is not acceptable on the data path, especially when the i2c
controller FIFO is small.

For sensor command FIFO_READ and CMD_DATA, calculate the exact
expected response length.

BUG=chromium:574480
TEST=check when idle (screen on), the interrupt rate from the i2c
controller is lower from ~500 to ~200.
Each FIFO_READ is 20/28 bytes when not in FIFO mode (instead of 120).
TEST=Check reading sensor directly also work as expected.
TEST=Check CTS tests android.hardware.cts.SingleSensorTests and
android.hardware.cts.SensorBatchingTests pass.

Change-Id: Idabc5f7b9f2c7ef29e7c95c03373828ff82d1cb9
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/320595
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>