UPSTREAM: Bluetooth: Add device shutdown routine for Intel Bluetooth device
This patch adds the device shutdown routine for Intel Bluetooth device.

Some platforms have BT LED issue with Intel Bluetooth device that BT LED goes
off 5 seconds after BT is turned off

For Intel Bluetooth device, the BT LED is turned off when:
- there is no active connection or radio activity
- USB is suspend

So, when the BT is turned off, it takes 5 seconds because USB suspend timeone
is 5 seconds by default. And if the USB suspend is not enabled, BT LED won't be
turned off.

To fix this issue, recently Intel Bluetooth firmware patch had been submitted
to turn off the BT LED immediately by the vendor specific command(0xFC3F). And
this patch sends this command to the device before closing the device.

For backward compatibility of this command  with old firmware, this command was
supported before, but it behaves same as HCI_RESET internally. So, it won't be
the problem even if the system doesn't have the latest firmware patch.

BUG=chrome-os-partner:32352
TEST=ran basic test to confirm the vendor specific command is sent
before closing the device.

Change-Id: Ifc76a5b457ecc8072adf2e4b8f442d9cd264fe23
Reviewed-on: https://chromium-review.googlesource.com/323963
Reviewed-by: Ethan Ge <ethan.ge@intel.com>
Commit-Queue: Ethan Ge <ethan.ge@intel.com>
Tested-by: Ethan Ge <ethan.ge@intel.com>
Tested-by: Eric Yang <eric.yang@intel.com>
Reviewed-by: Bowgo Tsai <bowgotsai@google.com>