Revert "PM / Suspend: Print wall time at suspend entry and exit"
This reverts commit 195382b8828f ("PM / Suspend: Print wall time at
suspend entry and exit").

Builds fail if SUSPEND is enabled and RTC_CLASS is disabled.

kernel/built-in.o: In function `rtc_time_to_tm':
	include/linux/rtc.h:32: undefined reference to `rtc_time64_to_tm'

Original commit is broken, it depends on RTC_CLASS but adding an
actual Kconfig depends is really heavy handed. We don't rely on these
markers today so we can revert the patch for now.

BUG=chromium:583462
TEST=None

Change-Id: Ia1a1c6ebaaa6cd82144176ffcde31842b7a78726
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/325552
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>