dm mpath: convert from spinlock to RCU for most locking
Most fast path code no longer requires m->lock.

Path switching requires m->lock but only for the callout to the
path-selector's ->select_path hook which requires the caller provide
locking (otherwise list corruption occurs in the path-selector).

choose_pgpath() is modified to _not_ update paths->current_pgpath unless
it also switches the path-group.  This is done to avoid needing to take
the write-side of the RCU everytime __multipath_map() calls
choose_pgpath().  In practice, repeat_count is typically very low
(e.g. 1) anyway -- to promote load balancing -- so choose_pgpath() is
called very frequently.  paths->current_pgpath will be reset if it is
failed via fail_path().

If repeat_count is > 1 then this strategy is ill-conceived because
paths->current_pgpath won't be changed; and as a result the path switch
once the repeat_count expires will only last for a single IO.
repeat_count was primarily for bio-based DM multipath's benefit.  There
is really no need for it anymore now that DM multipath is request-based.
As such, repeat_count > 1 is no longer honored and a warning is
displayed if the user attempts to use a value > 1.

Signed-off-by: Mike Snitzer <snitzer@redhat.com>